LI12425

Description

Tipos de dados e funções auxiliares para a realização do projeto de LI1 em 2024/25.

Tipos de dados

Básicos

type Creditos = Int

Moeda do jogo.

data Direcao

Direção de movimento de uma entidade no jogo.

Constructors

Norte

Sul

Este

Oeste

type Distancia = Float

Distância entre duas posições.

data Duracao

Representa uma duração em segundos

Constructors

Finita Tempo
Duração em segundos

Infinita
Duração infinita

type Posicao = (Float, Float)

Coordenada bilateral de uma entidade no jogo, representante do seu centro. O referencial tem origem no canto superior esquerdo, com eixos x e y positivos para a direita e para baixo, respectivamente.

type Semente = Int

Valor inicial que determina a sequência de números pseudo-aleatórios.

type Tempo = Float

Tempo em segundos.

Mapas

type Mapa = [[Terreno]]

Mapa do jogo composto por uma matriz de terrenos.

data Terreno

Tipo de terrenno do mapa.

Constructors

Relva
Torres constroem-se sobre o relvado do mapa.

Terra
A base e os portais constroem-se sobre caminhos de terra do mapa. Além disso, inimigos movem-se sobre estes terrenos.

Agua
Água para efeito decorativo, mas onde não se pode construir, nem os inimigos se podem mover.

Entidades

data Base

Base de operações do jogador.

Constructors

Base

Fields

vidaBase :: Float

Vida da base. Quando esta chega a zero, o jogador perde o jogo.

posicaoBase :: Posicao

Posição da base no mapa. A base deve estar sobre um terreno de terra.

creditosBase :: Creditos

Balanço de créditos do jogador.

data Torre

Torre que dispara projéteis contra inimigos.

Constructors

Torre

Fields

posicaoTorre :: Posicao

Posição da torre no mapa.

danoTorre :: Float

Redução de vida no inimigo pelo impacto do projétil.

alcanceTorre :: Float

Alcance circular da torre.

rajadaTorre :: Int

Número de máximo de inimigos simultaneamente atingidos por uma rajada de tiros.

cicloTorre :: Tempo

Ciclo de tempo entre rajadas de tiros.

tempoTorre :: Tempo

Tempo restante para a próxima rajada de tiros.

projetilTorre :: Projetil

Efeito secundário associado ao tipo de projétil da torre.

data Portal

Portal de entrada de inimigos no mapa.

Constructors

Portal

Fields

posicaoPortal :: Posicao

Posição do portal no mapa. O portal deve estar sobre um terreno de terra.

ondasPortal :: [Onda]

Ondas de inimigos que saem do portal.

data Inimigo

Inimigo que se move em direção à base do jogador.

Constructors

Inimigo

Fields

tipoInimigo :: TipoInimigo

posicaoInimigo :: Posicao

Posição do inimigo no mapa.

direcaoInimigo :: Direcao

Direção do último movimento do inimigo.

vidaInimigo :: Float

Vida do inimigo.

velocidadeInimigo :: Float

Velocidade do inimigo.

ataqueInimigo :: Float

Dano causado pelo inimigo na base do jogador.

butimInimigo :: Creditos

Créditos que o jogador recebe ao derrotar o inimigo.

projeteisInimigo :: [Projetil]

Efeitos secundários ativos no inimigo.

animacao :: Animacao

Animação selecionada

frameInimigo :: Int

Frame atual da animação do inimigo

tempoAInimigo :: Float

Tempo da animação

caminhoInimigo :: [Posicao]

Caminho a seguir até à torre

posMedia :: Maybe PosicaoIntermedia

Interpolacao

data TipoProjetil

Tipo de projétil disparado por uma torre.

Constructors

Fogo

Gelo

Resina

data Projetil

Projétil aplicado por uma torre.

Constructors

Projetil

Fields

tipoProjetil :: TipoProjetil

Tipo de projétil.

duracaoProjetil :: Duracao

Duração do efeito do projétil no inimigo.

Jogo

data Jogo

Estado do jogo. Um jogo é composto pela base, vários portais, várias torres, um mapa, vários inimigos e a loja.

Constructors

Jogo

Fields

baseJogo :: Base

Base de operações do jogador.

portaisJogo :: [Portal]

Portais de entrada de inimigos no mapa.

torresJogo :: [Torre]

Torres construídas pelo jogador.

mapaJogo :: Mapa

Mapa retangular do jogo.

inimigosJogo :: [Inimigo]

Inimigos em movimento no mapa.

lojaJogo :: Loja

Loja de torres disponíveis para construir.

data Onda

Onda de inimigos que saem de um portal.

Constructors

Onda

Fields

inimigosOnda :: [Inimigo]

Inimigos que compõem a onda.

cicloOnda :: Tempo

Tempo em segundos entre a entrada de cada inimigo.

tempoOnda :: Tempo

Tempo restante, em segundos, para a entrada do próximo inimigo da onda.

entradaOnda :: Tempo

Tempo restante, em segundos, para a entrada da onda.

type Loja = [(Creditos, Torre)]

Loja de torres disponíveis para construir por uma quantidade de créditos.

Outros

data Animacao

Animações possíveis

Constructors

Walk
Animação a andar

Frozen
Animação quando está congelado

data TipoInimigo

Tipos de Inimigos

Constructors

Normal
Inimigo padrão

Rapido
Inimigo agil mas com pouca vida e ataque

Tanque
Inimigo lento mas com muita vida e ataque

type PosicaoIntermedia = (Posicao, Posicao, Float)

Posicao intermedia entre duas posicoes de um inimigo

Funções auxiliares

geraAleatorios :: Semente -> Int -> [Int]

Função que gera uma lista de números aleatórios a partir de uma Semente.

Exemplos

>>> geraAleatorios 2425 3
[9108974057934916489,3509742222561512871,1534041518507426227]

>>> geraAleatorios 10 1
[3575835729477015470]