Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2024/25.

Documentation

validaJogo :: Jogo -> Bool

Verifica se o estado de um jogo é válido

validaBase :: Jogo -> Base -> Bool

Verifica se a Base é válida

validaSobreposicaoBase :: [Torre] -> [Portal] -> Base -> Bool

Verifica se a Base não está sobreposta a uma torre ou portal

validaBaseSobreTerraCredito :: [[(Terreno, (Int, Int))]] -> Base -> Bool

Verifica se a Base está colocada sobre terra e não tem crédito negativo

validaTorres :: Jogo -> [Torre] -> Bool

Verifica se as Torres são válidas

validaCicloTorre :: [Torre] -> Bool

Verifica se todas as Torres possuem ciclos de disparo não negativos

validaSobreposicaoTorres :: [Torre] -> Bool

Verifica se as Torres não estão sobrepostas

validaSobreRelvaAlcanceRajada :: [[(Terreno, (Int, Int))]] -> [Torre] -> Bool

Verifica se as Torres estão sobre relva e se o seu alcance é um valor positivo

validaInimigos :: Jogo -> [Portal] -> [Inimigo] -> [Torre] -> Bool

Verifica se os Inimigos são válidos

validaProjeteis :: [Projetil] -> Bool

Verifica se a lista de projéteis ativos de um Inimigo encontra-se normalizada

validaPosicaoPortal :: [Portal] -> Bool

Verifica se todos os Inimigos por lançar têm a posição do respetivo portal, nível de vida positivo, e lista de projéteis ativos vazia

inimigoSobreTerra :: [[(Terreno, (Int, Int))]] -> [Inimigo] -> Bool

Verifica se todos os Inimigos em jogo encontram-se sobre terra

sobreposicaoTorre :: [[(Terreno, (Int, Int))]] -> [Inimigo] -> [Torre] -> Bool

Verifica se os Inimigos não estão sobrepostos a Torres

validaPortais :: [Portal] -> Mapa -> Base -> [Torre] -> Bool

Verifica se os Portais são válidos

portalSobreTerra :: Mapa -> Portal -> Bool

Verifica se um Portal está posicionado sobre terra

existeCaminho :: Base -> [Portal] -> Mapa -> Bool

Verifica se existe um caminho de terra conectando pelo menos um Portal à Base

naoSobrepostos :: [Portal] -> Base -> [Torre] -> Bool

Verifica se os Portais não estão sobrepostos à Base ou às Torres

ondaAtivaPorPortal :: Portal -> Bool

Verifica se há, no máximo, uma onda ativa em um portal.

verificaOndasAtivas :: [Portal] -> Bool

Verifica se todos os portais possuem no máximo uma onda ativa.

getTerreno :: Mapa -> (Int, Int) -> Maybe Terreno

Obtém o terreno do mapa em uma posição específica

conectado :: Int -> Int -> Int -> Int -> Mapa -> Bool

Verifica se duas posições estão conectadas no mapa.

transformarMapa :: [[Terreno]] -> Int -> [[(Terreno, (Int, Int))]]

Transforma um mapa de terrenos em um mapa de terrenos com coordenadas

transformarMapaL :: [Terreno] -> Int -> Int -> [(Terreno, (Int, Int))]

Transforma uma linha de terrenos em uma linha de terrenos com coordenadas

acharBloco :: [[(Terreno, (Int, Int))]] -> (Float, Float) -> Maybe Terreno

Procura um terreno em uma matriz de blocos pelas coordenadas fornecidas

acharBlocoLinha :: [(Terreno, (Int, Int))] -> (Float, Float) -> Maybe Terreno

Procura um terreno em uma linha de blocos pelas coordenadas fornecidas