Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2024/25.

Documentation

inimigosNoAlcance :: Torre -> [Inimigo] -> [Inimigo]

Identifica quais Inimigos estão dentro do alcance de uma Torre

distancia :: Posicao -> Posicao -> Float

Calcula a distância euclidiana entre duas posições

atingeInimigo :: Torre -> Inimigo -> Inimigo

Função que aplica os efeitos de um ataque da Torre a um Inimigo

atualizarProjeteisInimigo :: [Projetil] -> Projetil -> [Projetil]

Função para atualizar a lista de projéteis do inimigo

somaDuracao :: Duracao -> Duracao -> Duracao

Função que soma as durações de projéteis

dobraDuracao :: Duracao -> Duracao

Função que duplica a duração de um projétil

ativaInimigo :: Tempo -> Portal -> [Inimigo] -> (Portal, [Inimigo])

Responsável por ativar Inimigos de um Portal, transferindo Inimigos de uma onda do Portal para a lista de Inimigos ativos

terminouJogo :: Jogo -> Bool

Decide se o jogo terminou

ganhouJogo :: Jogo -> Bool

Verifica se o jogador ganhou o jogo

perdeuJogo :: Jogo -> Bool

Verifica se o jogador perdeu o jogo