Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2024/25.

Documentation

atualizaJogo :: Tempo -> Jogo -> Jogo

Função principal de atualização do estado do jogo

atualizaTorres :: Tempo -> Jogo -> Jogo

Atualiza o estado das Torres (incluindo ataques a Inimigos)

atualizaTempoTorre :: Tempo -> Torre -> Torre

Atualiza o tempo restante de recarga de cada Torre

disparaTorre :: Torre -> [Inimigo] -> [Inimigo]

Função que dispara a torre, aplicando as regras de rajada e ciclo de tempo

atualizaCreditos :: [Inimigo] -> Base -> Base

Atualiza os créditos da base com base no butim dos inimigos derrotados

atualizaInimigos :: Tempo -> Jogo -> Jogo

Atualiza o estado dos inimigos e os créditos da base

atualizaInimigo :: Tempo -> Mapa -> [Torre] -> Posicao -> Inimigo -> Inimigo

Atualiza o estado de um inimigo com base no tempo, mapa, torres e posição da base

moveInimigo :: Tempo -> Mapa -> [Torre] -> Inimigo -> Inimigo

Move o inimigo com base no tempo, mapa, torres e efeitos dos projéteis, ajustando posição e direção

calculaProximoBloco :: Mapa -> Posicao -> Direcao -> Direcao

Calcula a nova direção para o inimigo

chegouBase :: Posicao -> Inimigo -> Bool

Determina se um inimigo alcançou a base

danoSeNaBase :: Posicao -> Inimigo -> Float

Determina o dano que um inimigo causa à base

encontraBase :: Posicao -> Posicao -> Bool

Verifica se duas posições coincidem

escolheNovaDirecao :: Mapa -> Posicao -> Direcao -> Direcao

Escolhe uma nova direção para o inimigo com base no mapa e sua posição atual

acharTerreno :: Mapa -> (Int, Int) -> Maybe Terreno

Obtém o tipo de terreno de uma posição no mapa, retornando Nothing se estiver fora dos limites

proximaPosicao :: Posicao -> Direcao -> Posicao

Calcula a próxima posição com base na direção

direcoesValidas :: Direcao -> [Direcao]

Retorna as direções válidas, excluindo a direção oposta

atualizaPortais :: Tempo -> Jogo -> Jogo

Lança novos Inimigos dos Portais

lancarInimigos :: Tempo -> Portal -> (Portal, [Inimigo])

Função que lança inimigos a partir de um portal com base no tempo