Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2024/25.

Documentation

validaJogo

Arguments

:: Jogo
Jogo a validar.

-> Bool
Resultado na forma True caso o jogo é válido ou False caso contrário.

A função validaJogo verifica se um dado jogo é ou não válido baseando-se numa série de condições verificadas por outras funções.

Exemplos

>>> validaJogo (Jogo (Base 100 (0,0) 100) [Portal (1,0) [Onda [Inimigo (1,0) Norte 100 1 5 5 [] [] 100 1] 5 1 (-5)]] [] mapa01 [] [] 1 (0,0) 0 (False,5,[]))
True

>>> validaJogo (Jogo (Base 100 (0,0) 100) [Portal (3,0) [Onda [Inimigo (1,0) Norte 100 1 5 5 [] [] 100 1] 5 1 (-5)]] [] mapa01 [] [] 1 (0,0) 0 (False,5,[]))
False

mapa01 :: Mapa

mapa01 é um exemplo de um mapa.

dist

Arguments

:: Posicao
Posição do primeiro ponto.

-> Posicao
Posição do segundo ponto.

-> Float
Distância entre os dois pontos.

A função dist determina a distância entre dois pontos de um plano.

Exemplos

>>> dist (0,0) (4,3)
5

>>> dist (2,2) (2,3)
1

encontraPontosAdjacentes

Arguments

:: Posicao
Ponto de foco.

-> [Posicao]
Lista de pontos.

-> [Posicao]
Pontos (não diagonalmente) adjacentes ao ponto de foco.

A função encontraPontosAdjacentes filtra, de uma lista de pontos, apenas aqueles que se encontram adjacentes (não diagonalmente) a um outro ponto de um plano.

Exemplos

>>> encontraPontosAdjacentes (4,4) [(5,5),(3,5),(3,3)]
[]

>>> encontraPontosAdjacentes (2,3) [(3,3),(3,4),(3,5)]
[(3,3)]

>>> encontraPontosAdjacentes (2,2) [(2,3),(2,1),(1,2)]
[(2,3),(2,1),(1,2)]

filtraPosicoesTerra

Arguments

:: Mapa
Mapa do jogo.

-> Posicao
Posição do ponto na matriz (deve ser (0,0) inicialmente para determinar corretamente as coordenadas).

-> [Posicao]
Posições correpondentes ao terreno de Terra.

A função filtraPosicoesTerra determina as coordenadas de pontos correnpondentes apenas ao terreno de Terra de um mapa.

Exemplo

>>> filtraPosicoesTerra mapa01 (0,0)
[(0.0,0.0),(1.0,0.0),(4.0,0.0),(1.0,1.0),(4.0,1.0),(1.0,2.0),(5.0,2.0),(1.0,3.0),(2.0,3.0),(5.0,3.0),(1.0,4.0),(2.0,4.0),(3.0,4.0),(4.0,4.0),(5.0,4.0)]

filtraPosicoesRelva

Arguments

:: Mapa
Mapa do jogo.

-> Posicao
Posição do ponto na matriz (deve ser (0,0) inicialmente para determinar corretamente as coordenadas).

-> [Posicao]
Posições correpondentes ao terreno de Relva.

A função filtraPosicoesRelva determina as coordenadas de pontos correnpondentes apenas ao terreno de Relva de um mapa.

Exemplo

>>> filtraPosicoesRelva mapa01 (0,0)
[(2.0,0.0),(0.0,1.0),(2.0,1.0),(5.0,1.0),(0.0,2.0),(2.0,2.0),(0.0,3.0),(4.0,3.0),(0.0,4.0),(4.0,5.0),(5.0,5.0)]

removeElemento

Arguments

:: Eq a

=> a
Elemento a remover.

-> [a]
Lista inicial.

-> [a]
Lista resultante.

A função removeElemento remove as ocorrências de um elemento específico de uma lista.

Exemplos

>>> removeElemento 3 [1,2,3,4,5]
[1,2,4,5]

>>> removeElemento 6 [1,2,3,4,5]
[1,2,3,4,5]

removeElementosRepetidos

Arguments

:: Eq a

=> [a]
Lista inicial.

-> [a]
Lista resultante.

A função removeElementosRepetidos remove todos os elementos repetidos numa lista.

Exemplos

>>> removeElementosRepetidos [1,2,3,4,4,5,5]
[1,2,3,4,5]

>>> removeElementosRepetidos [1,2,3,4,5]
[1,2,3,4,5]

verifExistPortais

Arguments

:: [Portal]
Lista dos portais.

-> Bool
Resultado na forma True caso existe pelo menos um portal ou False caso contrário.

(1.a) A função verifExistPortais verifica se existe pelo menos um portal durante o jogo.

Exemplos

>>> verifExistPortais []
False

>>> verifExistPortais [Portal (2,2) []]
True

verifPortaisSobreTerra

Arguments

:: [Posicao]
Posições dos portais.

-> Mapa
Mapa do jogo.

-> Bool
Resultado na forma True se todos os portais estão no caminho do terra ou False caso contrário.

(1.b) A função verifPortaisSobreTerra verifica se todos os portais do jogo estão posicionados sobre terra.

Exemplos

>>> verifPortaisSobreTerra [(1,0)] mapa01
True

>>> verifPortaisSobreTerra [(2,0)] mapa01
False

>>> verifPortaisSobreTerra [(1,0),(2,0)] mapa01
False

verifConexaoBasePortais

Arguments

:: Posicao
Posição da base.

-> [Posicao]
Posições dos portais.

-> [Posicao]
Posições dos caminhos de terra.

-> Bool
Resultado na forma True se existe algum caminho (de terra) da base a um dos portais ou False se não existe tal caminho.

(1.c) A função verifConexaoBasePortais verifica se existe, começando na base, um caminho de terra até a pelo menos um portal.

Exemplos

>>> verifConexaoBasePortais (0,0) [(2,3)] (filtraPosicoesTerra mapa01 (0,0))
True

>>> verifConexaoBasePortais (0,0) [(4,0)] (filtraPosicoesTerra mapa01 (0,0))
False

>>> verifConexaoBasePortais (0,0) [(2,3),(4,0)] (filtraPosicoesTerra mapa01 (0,0))
False

posicoesPossiveisParaPortais

Arguments

:: [Posicao]
Posições de terra do mapa.

-> Posicao
Posição da base.

-> [Posicao]
Posições possíveis para portais.

A função posicoesPossiveisParaPortais determina todas as posições que um portal pode ocupar num mapa, isto é, apenas as posições correspondentes a caminhos de terra que estão ligados à base.

Exemplos

>>> posicoesPossiveisParaPortais (filtraPosicoesTerra mapa01 (0,0)) (0,0)
[(0.0,0.0),(1.0,0.0),(1.0,1.0),(1.0,2.0),(1.0,3.0),(2.0,3.0),(2.0,4.0),(1.0,4.0),(3.0,4.0),(4.0,4.0),(5.0,4.0),(5.0,3.0),(5.0,2.0),(1.0,4.0),(2.0,4.0),(2.0,3.0),(3.0,4.0),(4.0,4.0),(5.0,4.0),(5.0,3.0),(5.0,2.0)]

>>> posicoesPossiveisParaPortais (filtraPosicoesTerra mapa01 (0,0)) (4,0)
[(4.0,0.0),(4.0,1.0)]

verifSobreposPortaisTorresBase

Arguments

:: [Posicao]
Posições dos portais.

-> [Posicao]
Posições das torres.

-> Posicao
Posição da base.

-> Bool
Resultado na forma True se não há qualquer sobreposição ou False se há alguma sobreposição.

(1.d) A função verifSobreposPortaisTorresBase verifica se há sobreposição dos portais com as torres ou com a base.

Exemplos

>>> verifSobreposPortaisTorresBase [(0,0),(3,3)] [(1,1),(2,2)] (3,4)
True

>>> verifSobreposPortaisTorresBase [(0,0),(3,3)] [(1,1),(2,2)] (3,3)
False

>>> verifSobreposPortaisTorresBase [(0,0)] [(1,1),(2,2)] (3,3)
True

>>> verifSobreposPortaisTorresBase [(0,0),(3,3)] [(1,1),(2,2),(0,0)] (3,4)
False

>>> verifSobreposPortaisTorresBase [(3,3)] [(1,1),(2,2)] (3,4)
True

>>> verifSobreposPortaisTorresBase [(0,0),(3,3)] [(1,1),(3,3)] (0,0)
False

verifOndasAtivasPorPortal

Arguments

:: Portal
Portal de escolha.

-> Bool
Resultado na forma True se não há mais que uma onda ativa para o portal escolhido ou False caso contrário.

(1.e) A função verifOndasAtivasPorPortal verifica se há, no máximo, uma onda ativa para um dado portal.

Exemplos

>>> verifOndasAtivasPorPortal (Portal (0,0) [Onda [Inimigo (0,0) Norte 5 5 5 5 [] [] 5 1] 5 5 5,Onda [Inimigo (1,1) Este 10 10 10 10 [] [] 10 1] 10 10 10])
True

>>> verifOndasAtivasPorPortal (Portal (0,0) [Onda [] 5 5 (-5),Onda [] 10 10 (-10)])
True

>>> verifOndasAtivasPorPortal (Portal (0,0) [Onda [Inimigo (0,0) Norte 5 5 5 5 [] [] 5 1] 5 5 (-5),Onda [] 10 10 (-10)]) ~=? True
True

>>> verifOndasAtivasPorPortal (Portal (0,0) [Onda [Inimigo (0,0) Norte 5 5 5 5 [] [] 5 1] 5 5 (-5),Onda [Inimigo (1,1) Este 10 10 10 10 [] [] 10 1] 10 10 (-10)])
False

verifInimigosInativos

Arguments

:: Portal
Portal do qual os inimigos irão sair.

-> Bool
Resultado na forma True se todas as condições são verificadas ou False se pelo menos uma não está correta.

(2.a) A função verifInimigosInativos verifica se os inimigos inativos (ainda por sair do portal) estão nas condições corretas, isto é, estão na posição do seu portal, possuem vida positiva e não possuem projéteis ativos.

Exemplos

>>> verifInimigosInativos (Portal (2,2) [])
True

>>> verifInimigosInativos (Portal (2,2) [Onda [Inimigo (2,2) Norte 1 1 1 1 [] [] 1 1] 1 1 1])
True

>>> verifInimigosInativos (Portal (2,2) [Onda [Inimigo (2,3) Norte 1 1 1 1 [] [] 1 1] 1 1 1])
False

>>> verifInimigosInativos (Portal (2,2) [Onda [Inimigo (2,2) Norte 0 1 1 1 [] [] 1 1] 1 1 1])
False

>>> verifInimigosInativos (Portal (2,2) [Onda [Inimigo (2,2) Norte 1 1 1 1 [Projetil Fogo (Finita 1)] [] 1 1] 1 1 1])
False

verifPosicaoInimigoPortal

Arguments

:: [[Inimigo]]
Inimigos do portal.

-> Posicao
Posição do portal.

-> Bool
Resultado na forma True caso todos os inimigos estão na posição do seu portal ou False caso contrário.

A função verifPosicaoInimigoPortal verifica se as posições dos inimigos de um portal coincidem com a posição do próprio portal.

Exemplos

>>> verifPosicaoInimigoPortal [[Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1]] (1,1)
True

>>> verifPosicaoInimigoPortal [[Inimigo (1,2) Norte 1 1 1 1 [] [] 1 1]] (1,1)
False

>>> verifPosicaoInimigoPortal [[Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1],[Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1]] (1,1)
False

vidaInimigoPositiva

Arguments

:: [[Inimigo]]
Inimigos do portal.

-> Bool
Resultado na forma True se nenhum inimigo possui vida não positiva ou False se pelo menos um não respeita a condição.

A função vidaInimigoPositiva verifica se todos os inimigos de um portal possuem vida positiva.

Exemplos

>>> vidaInimigoPositiva [[Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1]]
True

>>> vidaInimigoPositiva [[Inimigo (1,1) Norte (-1) 1 1 1 [] [] 1 1]] ~=? False
False

>>> vidaInimigoPositiva [[Inimigo (1,1) Norte (-1) 1 1 1 [] [] 1 1],[Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1]]
False

listaProjeteisAtivosInimigos

Arguments

:: [[Inimigo]]
Inimigos do portal.

-> Bool
Resultado na forma True se nenhum inimgo possui projéteis ativos ou False caso contrário.

A função listaProjeteisAtivosInimigos verifica se nenhum inimigo num portal possui projéteis ativos.

Exemplos

>>> listaProjeteisAtivosInimigos [[Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1]]
True

>>> listaProjeteisAtivosInimigos [[Inimigo (1,1) Norte 1 1 1 1 [Projetil Fogo (Finita 1)] [] 1 1]]
False

>>> listaProjeteisAtivosInimigos [[Inimigo (1,1) Norte 1 1 1 1 [Projetil Fogo (Finita 1)] [] 1 1],[Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1]]
False

todosInimigosSobreTerra

Arguments

:: [Posicao]
Posições dos inimigos.

-> [Posicao]
Posições de terreno de terra.

-> Bool
Resultado na forma True se todos os inimigos estão posicionados sobre terra ou False caso contrário.

(2.b) A função todosInimigosSobreTerra verifica se todos os inimigos ativos no jogo estão posicionados sobre terreno de terra.

Exemplos

>>> todosInimigosSobreTerra [(1,1),(2,3)] (filtraPosicoesTerra mapa01 (0,0))
True

>>> todosInimigosSobreTerra [(1,1),(2,3),(2,0)] (filtraPosicoesTerra mapa01 (0,0))
False

inimigosSobreposTorres

Arguments

:: [Inimigo]
Lista de inimigos.

-> [Torre]
Lista de torres.

-> Bool
Resultado na forma True se nenhum inimigo se encontra posicionado sobre uma torre ou False caso contrário.

(2.c) A função inimigosSobreposTorres verifica se nenhum inimigo se encontra posicionado sobre uma torre.

Exemplos

>>> inimigosSobreposTorres [Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1,Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1] [Torre (2,0) 1 1 1 1 1 (Projetil Gelo Infinita) 1 1]
True

>>> inimigosSobreposTorres [Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1,Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1] [Torre (2,2) 1 1 1 1 1 (Projetil Gelo Infinita) 1 1]
False

>>> inimigosSobreposTorres [Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1,Inimigo (2,0) Norte 2 2 2 2 [] [] 2 1] [Torre (2,0) 1 1 1 1 1 (Projetil Gelo Infinita) 1 1]
False

velocidadeInimigosNaoNegativa

Arguments

:: [Inimigo]
Lista de Inimigos

-> Bool
Resultado na forma True caso todos os inimigos possuem velocidade não negativa ou False caso contrário.

(2.d) A função velocidadeInimigosNaoNegativa verifica se nenhum inimigo possui velocidade negativa.

Exemplos

>>> velocidadeInimigosNaoNegativa [Inimigo (1,1) Norte 1 1 1 1 [] [] 1 1,Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1]
True

>>> velocidadeInimigosNaoNegativa [Inimigo (1,1) Norte 1 0 1 1 [] [] 1 1,Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1]
True

>>> velocidadeInimigosNaoNegativa [Inimigo (1,1) Norte 1 (-1) 1 1 [] [] 1 1,Inimigo (2,2) Norte 2 2 2 2 [] [] 2 1]
False

verifProjeteisNormalizados

Arguments

:: [Projetil]
Lista de projéteis.

-> Bool
Resultado na forma True se os projéteis estão normalizados ou False caso contrário.

(2.e) A função verifProjeteisNormalizados verifica se os projéteis de uma lista estão normalizados, ou seja, não pode haver vários projéteis do mesmo tipo e não pode haver, simultâneamente, projéteis de gelo e fogo nem de resina e fogo.

Exemplos

>>> verifProjeteisNormalizados [Projetil Fogo Infinita]
True

>>> verifProjeteisNormalizados [Projetil Fogo Infinita,Projetil Fogo (Finita 1)]
False

>>> verifProjeteisNormalizados [Projetil Fogo Infinita,Projetil Gelo (Finita 1)]
False

>>> verifProjeteisNormalizados [Projetil Fogo Infinita,Projetil Resina Infinita]
False

>>> verifProjeteisNormalizados [Projetil Gelo Infinita,Projetil Resina Infinita]
True

>>> verifProjeteisNormalizados [Projetil Gelo Infinita,Projetil Resina Infinita,Projetil Fraqueza Infinita]
True

verifTorresSobreRelva

Arguments

:: [Torre]
Lista de torres.

-> Mapa
Mapa do jogo.

-> Bool
Resultado na forma True caso todas as torres estão posicionadas sobre relva ou False caso contrário.

(3.a) A função verifTorresSobreRelva verifica se todas as torres estão posicionadas em terreno de relva.

Exemplos

>>> verifTorresSobreRelva [Torre (2,0) 1 1 1 1 1 (Projetil Gelo Infinita) 1 1] mapa01
True

>>> verifTorresSobreRelva [Torre (1,0) 1 1 1 1 1 (Projetil Gelo Infinita) 1 1] mapa01
False

>>> verifTorresSobreRelva [Torre (2,0) 1 1 1 1 1 (Projetil Gelo Infinita) 1 1,Torre (1,0) 1 1 1 1 1 (Projetil Fogo Infinita) 1 1] mapa01
False

verifAlcanceTorres

Arguments

:: [Torre]
Lista de torres.

-> Bool
Resultado na forma True se nenhuma torre possui alcance negativo ou False se pelo menos uma possui alcance negativo.

(3.b) A função verifAlcanceTorres verifica se todas as torres possuem um alcance não negativo.

Exemplos

>>> verifAlcanceTorres [Torre (1,1) 1 1 1 1 1 (Projetil Resina Infinita) 1 1]
True

>>> verifAlcanceTorres [Torre (1,1) 1 0 1 1 1 (Projetil Resina Infinita) 1 1]
True

>>> verifAlcanceTorres [Torre (1,1) 1 (-1) 1 1 1 (Projetil Resina Infinita) 1 1]
False

verifRajadaTorres

Arguments

:: [Torre]
Lista de torres.

-> Bool
Resultado na forma True todas as torres possuem rajada positiva ou False se pelo menos uma possui rajada não positiva.

(3.c) A função verifRajadaTorres verifica se todas as torres possuem uma rajada positiva.

Exemplos

>>> verifRajadaTorres [Torre (1,1) 1 1 1 1 1 (Projetil Fraqueza Infinita) 1 1]
True

>>> verifRajadaTorres [Torre (1,1) 1 1 0 1 1 (Projetil Fraqueza Infinita) 1 1]
False

>>> verifRajadaTorres [Torre (1,1) 1 1 (-1) 1 1 (Projetil Fraqueza Infinita) 1 1]
False

verifCicloTorres

Arguments

:: [Torre]
Lista de torres.

-> Bool
Resultado na forma True se nenhuma torre possui ciclo negativo ou False se pelo menos uma possui ciclo negativo.

(3.d) A função verifCicloTorres verifica se todas as torres possuem um ciclo não negativo.

Exemplos

>>> verifCicloTorres [Torre (1,1) 1 1 1 1 1 (Projetil Resina Infinita) 1 1]
True

>>> verifCicloTorres [Torre (1,1) 1 1 1 0 1 (Projetil Resina Infinita) 1 1]
True

>>> verifCicloTorres [Torre (1,1) 1 1 1 (-1) 1 (Projetil Resina Infinita) 1 1]
False

verifSobreposTorres

Arguments

:: [Torre]
Lista de torres.

-> Bool
Resultado na forma True se não há sobreposição de torres ou False caso contrário.

(3.e) A função verifSobreposTorres verifica se as torres não estão sobrepostas.

Exemplos

>>> verifSobreposTorres [Torre (1,1) 1 1 1 1 1 (Projetil Resina Infinita) 1 1,Torre (2,2) 1 1 1 1 1 (Projetil Fogo Infinita) 1 1]
True

>>> verifSobreposTorres [Torre (1,1) 1 1 1 1 1 (Projetil Resina Infinita) 1 1,Torre (1,1) 1 1 1 1 1 (Projetil Fogo Infinita) 1 1]
False

verifBaseSobreTerra

Arguments

:: Base
Base do jogo.

-> Mapa
Mapa do jogo.

-> Bool
Resultado na forma True se a base está posicionada sobre terra ou False caso contrário.

(4.a) A função verifBaseSobreTerra verifica se a base está posicionada sobre terra.

Exemplos

>>> verifBaseSobreTerra (Base 1 (0,0) 1) mapa01
True

>>> verifBaseSobreTerra (Base 1 (2,2) 1) mapa01
False

verifCreditosBase

Arguments

:: Base
Base do jogo.

-> Bool
Resultado na forma True se os créditos da base não são negativos ou False se são negativos.

(4.b) A função verifCreditosBase verifica se os créditos da base não são negativos.

Exemplos

>>> verifCreditosBase (Base 1 (0,0) 1)
True

>>> verifCreditosBase (Base 1 (0,0) 0)
True

>>> verifCreditosBase (Base 1 (0,0) (-1))
False

verifSobreposBaseTorresPortais

Arguments

:: Posicao
Posição da base.

-> [Posicao]
Posições das torres.

-> [Posicao]
Posições dos portais.

-> Bool
Resultado na forma True caso não haja sobreposições ou False caso contrário.

(4.c) A função verifSobrepoisBaseTorresPortais verifica se a base não está sobreposta com torres ou com portais.

Exemplos

>>> verifSobreposBaseTorresPortais (1,1) [(0,0),(2,2)] [(3,3),(4,4)]
True

>>> verifSobreposBaseTorresPortais (1,1) [(0,0),(1,1)] [(3,3),(4,4)]
False

>>> verifSobreposBaseTorresPortais (1,1) [(0,0),(2,2)] [(1,1),(4,4)]
False

>>> verifSobreposBaseTorresPortais (1,1) [(1,1),(2,2)] [(1,1),(4,4)]
False