Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2024/25.

Documentation

validaJogo

Arguments

:: Jogo
recebe o jogo

-> Bool
devolve um Booleano

A função validaJogo é responsável por fazer todas a verificações necessárias para que um jogo seja válido.

Exemplos de Aplicação:

Expand

>>> validaJogo jogoEx
True

Funções Auxiliares:

Expand

Função validaPortais

Função validaInimigosAll

Função projeteisNorm

Função validaTorres

Função validaBase

validaPortais

Arguments

:: [Portal]
recebe a lista de portais em jogo

-> Mapa
recebe o mapa do jogo

-> Posicao
recebe a posição da base

-> [Torre]
recebe a lista de torres em jogo //todo: não devia verificar se portal sobrepoe portal?

-> Bool
devolve um Bool

A função validaPortais é responsável por fazer todas a verificações, relativas a portais, necessárias para que um jogo seja válido. Esta função é utilizada pela função validaJogo.

Exemplos de Aplicação:

Expand

>>> validaPortais [] mapa01 (3,2) torresEx
False

>>> validaPortais [(Portal (1,1) ondaEx)] mapa01 (3,4) torresEx
True

portaisTerra

Arguments

:: [Posicao]
recebe a lista de portais em jogo

-> Mapa
recebe o mapa do jogo

-> Bool
devolve um Booleano

A função portaisTerra é responável por verificar se todos os portais de uma lista de portais estão sobre o terreno de Terra. Utiliza a função emTerra.

Exemplos de Aplicação:

Expand

>>> portaisTerra [(1,1)] mapa01
True

emTerra

Arguments

:: Posicao
recebe a posição de um determinado bloco

-> Mapa
recebe o mapa do jogo

-> Bool
devolve um Booleano

A função emTerra é responável por verificar se um dado bloco de coordenada (x,y) é um terreno de Terra.

Exemplos de Aplicação:

Expand

>>> emTerra (1,1) mapa01
True

dentroMapa

Arguments

:: Mapa
recebe o mapa do jogo

-> Posicao
recebe a posição de um determinado bloco

-> Bool
devolve um Booleano

A função dentroMapa é responsável por verificar se uma bloco de coordenada (x,y) pertence ou não ao mapa de jogo

Exemplos de aplicação:

Expand

>>> dentroMapa mapa01 (0,0)
True

>>> dentroMapa mapa01 (20,54)
False

checkCaminho

Arguments

:: Mapa
recebe o mapa do jogo

-> Posicao
recebe a posição de um Portal

-> Posicao
recebe a posição da base

-> [Posicao]
recebe um acomulador inicialmente vazio que acomula os blocos já visitados

-> Bool
devolve um Booleano

A função checkCaminho é responsável por verificar para um dado mapa com um portal e uma base, se existe um caminho em terra que ligue o portal e a base. No caso

Exemplos de Aplicação:

Expand

>>> checkCaminho mapa01 (0,0) (4,4) []
True

>>> checkCaminho mapa01 (0,0) (4,5) []
False

portalSobrepoe

Arguments

:: Posicao
recebe o portal em questão

-> Posicao
recebe a base do jogo

-> [Posicao]
recebe as torres do jogo

-> Bool
devolve um Booleano

A função portalSobrepoe é responsável por verificar se um Portal sobrepõe a Base ou alguma torre

Exemplos de Aplicação:

Expand

>>> portalSobrepoe (0,0) (5,4) [(2,2)]
False

>>> portalSobrepoe (5,4) (5,4) [(2,2)]
True

umaOndaAtiva

Arguments

:: Portal
recebe um Portal do jogo

-> Bool
devolve um Booleano

A função umaOndaAtiva é responsável por verificar se um dado portal tem no máximo uma onda átiva.

Exemplos de aplicação:

Expand

>>> umaOndaAtiva (Portal (0,0) [(Onda [] 3 0 0), (Onda [inimigoEx] 5 5 0)])
True

validaInimigosAll

Arguments

:: [Portal]
recebe os portais em jogo

-> Mapa
recebe o mapa do jogo

-> [Torre]
recebe as torres em jogo

-> [Inimigo]
recebe os inimigos em jogo

-> Bool
devolve um Booleano

A função validaInimigosAll faz a validação dos inimigos de vários portais através da função validaInimigos

Exemplos de Aplicação:

Expand

>>> validaInimigosAll portalEx mapa01 torresEx [inimigoEx]
True

validaInimigos

Arguments

:: Portal
recebe um portal em jogo

-> Mapa
recebe o mapa do jogo

-> [Torre]
recebe as torres em jogo

-> [Inimigo]
recebe os inimigos spawnados

-> Bool
devolve um Booleano

A função validaInimigos é responsável por fazer todas a verificações, relativas a portais, necessárias para que um jogo seja válido. Esta função é utilizada pela função validaJogo.

Exemplos de Aplicação:

Expand

>>> validaInimigos (Portal (0,0) [Onda [Inimigo Normal (0,0) Sul 10 0.5 10 10 [] Walk 1 0 [] Nothing] 3 0 0]) mapa01 torresEx [inimigoEx]
True

inimigoNoPortal

Arguments

:: Posicao
recebe a posição de um portal

-> [Posicao]
recebe as posições dos respetivos inimigos

-> Bool
devolve um Booleano

A função inimigoNoPortal é responsável por verificar se um inimigo tem a posição do respetivo portal.

Exemplo de Aplicação:

Expand

>>> inimigoNoPortal (0,0) [(0,0)]
True

inimigoStatsOk

Arguments

:: [Inimigo]
recebe os inimigos que ainda não spawnaram

-> Bool
devolve um Booleano

A função inimigoStatsOk é responsável por verificar se as stats dos inimigos são válidas. Isto inclui: o nível de vida é positivo; a lista de projeteis ativos é vazia; a velocidade não é negativa.

Exemplos de Aplicação:

Expand

>>> inimigoStatsOk [inimigoEx]
False

>>> inimigoStatsOk [(Inimigo Normal (0,0) Sul 10 0.5 10 10 [] Walk 1 0 [] Nothing)]
True

checkVel

Arguments

:: [Inimigo]
recebe a lista de todos os inimigos do jogo

-> Bool
devolve um Booleano

A função checkVel é responsável por verificar se a velocidade de todos os inimigos é não negativa. É utilizada pela função validaInimigos.

projeteisNorm

Arguments

:: [Projetil]
recebe a lista de projeteis de um inimigo já em jogo

-> Bool
devolve um Booleano

A função projeteisNorm é responsável por verificar se a lista de projeteis de um inimigo já em jogo, está normalizada. A função checkSinergias é integrada como auxiliar

Exemplos de aplicação:

Expand

>>> projeteisNorm [(Projetil Fogo (Finita 3)), (Projetil Gelo (Finita 3))]
False

>>> projeteisNorm [(Projetil Gelo (Finita 3)), (Projetil Gelo (Finita 3))]
False

checkSinergias

Arguments

:: [Projetil]
recebe a lista de projeteis de um inimigo já em jogo

-> Bool
devolve um Booleano

A função checkSinergias é responsável por verificar as condições de compatibilidade do jogo. Nomeadamente, verifica se existem, simultaneamente, projéteis do tipo Fogo e Resina, ou do Tipo Fogo e Gelo.

Exemplos de Aplicação:

Expand

>>> checkSinergias [(Projetil Fogo (Finita 3)), (Projetil Gelo (Finita 3))]
False

>>> checkSinergias [(Projetil Resina (Finita 3)), (Projetil Gelo (Finita 3))]
True

inimigoSobrepoe

Arguments

:: Posicao
recebe a posição de um inimigo

-> [Torre]
recebe as torres do jogo

-> Bool
devolve um Booleano

A função inimigoSobrepoe é responsável por verificar se um inimigo sobrepõe alguma Torre

Exemplos de Aplicação:

Expand

>>> inimigoSobrepoe (2,2) torresEx
True

validaTorres

Arguments

:: [Torre]
recebe a lista de Torres em jogo

-> Mapa
recebe o mapa do jogo

-> Bool
devolve um Booleano

A função validaTorres é responsável por por fazer todas a verificações, relativas às Torres, necessárias para que um jogo seja válido. Esta função é utilizada pela função validaJogo.

Exemplos de Aplicação:

Expand

>>> validaTorres torresEx mapa01
True

emRelva

Arguments

:: Posicao
recebe a posição de um determinado bloco

-> Mapa
recebe o mapa do jogo

-> Bool
devolve um Booleano

A função emRelva é responável por verificar se um dado bloco de coordenada (x,y) é um terreno de Relva.

Exemplos de Aplicação:

Expand

>>> emRelva (2,0) mapa01
True

validaBase

Arguments

:: Base
recebe a base do jogo

-> Mapa
recebe o mapa do jogo

-> [Portal]
recebe os portais do jogo

-> [Torre]
recebe as torres do jogo

-> Bool
devolve um Booleano

A função validaBase é responsável por fazer todas a verificações, relativas à base, necessárias para que um jogo seja válido. Esta função é utilizada pela função validaJogo.

Exemplos de Aplicação:

Expand

>>> validaBase baseEx mapa01 portalEx torresEx
True

baseSobrepoe

Arguments

:: Base
recebe a base do jogo

-> [Portal]
recebe os portais do jogo

-> [Torre]
recebe as torres do jogo

-> Bool
devolve um Booleano

A função baseSobrepoe é responsável por verificar se a Base sobrepõe algum portal ou alguma torre

Exemplos de Aplicação:

Expand

>>> baseSobrepoe baseEx portalEx torresEx
False

mapa01 :: Mapa

Função para testes

jogoEx :: Jogo

Função para testes

lojaEx :: Loja

Função para testes

inimigoEx :: Inimigo

Função para testes

baseEx :: Base

Função para testes

torresEx :: [Torre]

Função para testes

portalEx :: [Portal]

Função para testes

ondaEx :: [Onda]

Função para testes