Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2024/25.

Documentation

validaJogo :: Jogo -> Bool

Verifica se um jogo é válido.

Exemplo

>>> validaJogo Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

existePortal :: Jogo -> Bool

Verefica se existe pelo menos um portal.

Exemplo

>>> existePortal Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

portaisSobreTerra :: Jogo -> Bool

Verifica se todos os portais estão sobre terra.

Exemplo

>>> portaisSobreTerra Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

posicaoSobreTerra :: Posicao -> Mapa -> Bool

Verifica se uma posição está sobre Terra.

Exemplo

>>> posicaoSobreTerra (1, 1) [[Terra, Terra, Terra], [Relva, Relva, Relva],[Terra, Terra, Terra]]
True

existeCaminho :: Posicao -> Posicao -> Mapa -> Bool

Verifica se existe um caminho de terra entre duas posições. | Verifica se existe um caminho de terra entre duas posições.

Exemplo

>>> existeCaminho (0, 0) (2, 2) [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]]
True

portaisNaoSobrepostos :: Jogo -> Bool

Verifica que o portal nao esta sobreposto a torres ou base.

Exemplo

>>> portaisNaoSobrepostos Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

sobreposicao :: Posicao -> Posicao -> Bool

Verifica se duas posições estão sobrepostas.

Exemplo

>>> sobreposicao (1, 1) (1, 1)
True

ondasAtivasValidas :: Jogo -> Bool

Verifica se há no máximo uma onda ativa por portal.

Exemplo

>>> ondasAtivasValidas Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

portaisValidos :: Jogo -> Bool

Verifica se todos os portais são válidos.

Exemplo

>>> portaisValidos Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

inimigosPorLancarValidos :: Jogo -> Bool

Verifica se todos os inimigos por lançar têm a posição do respetivo portal, nível de vida positivo e lista de projéteis ativos vazia.

Exemplo

>>> inimigosPorLancarValidos Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

inimigosEmJogoValidos :: Jogo -> Bool

Verifica se todos os inimigos em jogo estão sobre terra, não estão sobrepostos a torres e têm velocidade não negativa.

Exemplo

>>> inimigosEmJogoValidos Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

projeteisNormalizados :: [Projetil] -> Bool

Verifica se a lista de projéteis ativos está normalizada.

Exemplo

>>> projeteisNormalizados [Projetil {tipoProjetil = Fogo}, Projetil {tipoProjetil = Gelo}]
True

inimigosValidos :: Jogo -> Bool

Verifica se todos os inimigos são válidos.

Exemplo

>>> inimigosValidos Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

torresSobreRelva :: Jogo -> Bool

Verifica se todas as torres estão sobre relva.

Exemplo

>>> torresSobreRelva Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

posicaoSobreRelva :: Posicao -> Mapa -> Bool

Verifica se uma posição está sobre Relva.

Exemplo

>>> posicaoSobreRelva (1, 1) [[Terra, Terra, Terra], [Relva, Relva, Relva],[Terra, Terra, Terra]]
True

alcancePositivo :: Jogo -> Bool

Verifica se todas as torres têm alcance positivo.

Exemplo

>>> alcancePositivo Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

rajadaPositiva :: Jogo -> Bool

Verifica se todas as torres têm rajada positiva.

Exemplo

>>> rajadaPositiva Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

cicloNaoNegativo :: Jogo -> Bool

Verifica se todas as torres têm ciclo não negativo.

Exemplo

>>> cicloNaoNegativo Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

torresNaoSobrepostas :: Jogo -> Bool

Verifica se as torres não estão sobrepostas.

Exemplo

>>> torresNaoSobrepostas Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}, Torre {posicaoTorre = (1, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

torresValidas :: Jogo -> Bool

Verifica se todas as torres são válidas.

Exemplo

>>> torresValidas Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}, Torre {posicaoTorre = (1, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

baseSobreTerra :: Jogo -> Bool

Verifica se a base está colocada sobre terra.

Exemplo

>>> baseSobreTerra Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

baseCreditoPositivo :: Jogo -> Bool

Verifica se a base não tem crédito negativo.

Exemplo

>>> baseCreditoPositivo Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

baseNaoSobreposta :: Jogo -> Bool

Verifica se a base não está sobreposta a uma torre ou portal.

Exemplo

>>> baseNaoSobreposta Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True

baseValida :: Jogo -> Bool

Verifica se a base é válida.

Exemplos

>>> baseValida Jogo {mapaJogo = [[Terra, Terra, Terra], [Relva, Relva, Relva], [Terra, Terra, Terra]], portaisJogo = [Portal {posicaoPortal = (1, 1), ondasPortal = [Onda {tempoOnda = 10, inimigosOnda = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}]}]}], torresJogo = [Torre {posicaoTorre = (2, 2), alcanceTorre = 5, rajadaTorre = 3, cicloTorre = 2}], inimigosJogo = [Inimigo {posicaoInimigo = (1, 1), vidaInimigo = 100, projeteisInimigo = [], velocidadeInimigo = 1}], baseJogo = Base {posicaoBase = (0, 0), creditosBase = 100}}
True