Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2024/25.

Documentation

inimigoEstaNoAlcance

Arguments

:: Torre
Torre de escolha.

-> Inimigo
Inimigo de escolha.

-> Bool
Resultado na forma True caso o inimigo esteja no alcance da torre ou False caso contrário.

A função inimigoEstaNoAlcance verifica se um dado inimigo está no alcance de uma torre.

Exemplos

>>> inimigoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (5,4) Norte 50 50 50 50 [] [] 50 1)
True

>>> inimigoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1)
True

>>> inimigoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (9,4) Norte 50 50 50 50 [] [] 50 1)
False

>>> inimigoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (8,8) Norte 50 50 50 50 [] [] 50 1)
False

somaDuracoes

Arguments

:: Projetil
Primeiro projétil.

-> Projetil
Segundo projétil (do mesmo tipo que o primeiro).

-> Projetil
Projétil resultante.

A função somaDuracoes recebe dois projéteis e devolve um projétil cuja duração é a soma das durações dos projéteis iniciais, mantendo o seu tipo. Esta função apenas funciona corretamente se os projéteis são do mesmo tipo.

Exemplos

>>> somaDuracoes (Projetil Fogo (Finita 4)) (Projetil Fogo (Finita 5))
Projetil Fogo (Finita 9)

>>> somaDuracoes (Projetil Gelo Infinita) (Projetil Gelo (Finita 5))
Projetil Gelo Infinita

>>> somaDuracoes (Projetil Resina (Finita 5)) (Projetil Resina Infinita)
Projetil Resina Infinita

>>> somaDuracoes (Projetil Fraqueza Infinita) (Projetil Fraqueza Infinita)
Projetil Fraqueza Infinita

verifFraqueza

Arguments

:: Projetil
Projétil recebido.

-> Bool
Resultado no tipo True caso o projétil seja de fraqueza ou False caso contrário.

A função verifFraqueza verifica se um projétil é do tipo Fraqueza.

Exemplos

>>> verifFraqueza (Projetil Fraqueza Infinita)
True

>>> verifFraqueza (Projetil Fraqueza (Finita 3))
True

>>> verifFraqueza (Projetil Fogo Infinita)
False

verifNaoFraqueza

Arguments

:: Projetil
Projétil recebido.

-> Bool
Resultado no tipo False caso o projétil seja de fraqueza ou True caso contrário.

A função verifNaoFraqueza verifica se um projétil não é do tipo fraqueza.

Exemplos

>>> verifNaoFraqueza (Projetil Fraqueza Infinita)
False

>>> verifNaoFraqueza (Projetil Fraqueza (Finita 3))
False

>>> verifNaoFraqueza (Projetil Fogo Infinita)
True

normalizaProjeteis

Arguments

:: [Projetil]
Lista inicial de projéteis.

-> [Projetil]
Lista de projéteis normalizados.

A função normalizaProjeteis normaliza os projéteis de uma lista.

Exemplos

>>> normalizaProjeteis [Projetil Fraqueza (Finita 5),Projetil Fraqueza Infinita,Projetil Gelo Infinita,Projetil Fogo Infinita]
[Projetil Fraqueza Infinita]

>>> normalizaProjeteis [Projetil Fraqueza (Finita 5),Projetil Fraqueza (Finita 1),Projetil Resina Infinita,Projetil Fogo (Finita 4)]
[Projetil Fraqueza (Finita 6),Projetil Fogo (Finita 8)]

normalizaProjeteisNaoFraqueza

Arguments

:: [Projetil]
Lista inicial de projéteis (que não são de fraqueza).

-> [Projetil]
Lista de projéteis normalizados.

A função normalizaProjeteis normaliza os projéteis (que não são de fraqueza) de uma lista.

Exemplos

>>> normalizaProjeteisNaoFraqueza [Projetil Gelo Infinita,Projetil Fogo Infinita]
[]

>>> normalizaProjeteisNaoFraqueza [Projetil Resina (Finita 1),Projetil Fogo (Finita 3)]
[Projetil Fogo (Finita 6)]

novosProjeteisInimigo

Arguments

:: Projetil
Projétil que atinge o inimigo.

-> [Projetil]
Lista de projéteis do inimigo.

-> [Projetil]
Lista de projéteis resultante.

A função novosProjeteisInimigo calcula o que acontece à lista de projéteis de um inimigo quando atingido por outro projétil.

Exemplos

>>> novosProjeteisInimigo (Projetil Fogo Infinita) [Projetil Gelo Infinita]
[]

>>> novosProjeteisInimigo (Projetil Fraqueza (Finita 4)) [Projetil Resina Infinita,Projetil Fraqueza (Finita 6)]
[Projetil Fraqueza (Finita 10),Projetil Resina Infinita]

existOndaAtiva

Arguments

:: [Onda]
Lista de ondas.

-> Bool
Resultado na forma True se existe pelo menos uma onda ativa na lista ou False caso contrário.

A função existOndaAtiva verifica se exise pelo menos uma onda ativa numa lista de ondas.

Exemplos

>>> existOndaAtiva []
False

>>> existOndaAtiva [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) 5]
False

>>> existOndaAtiva [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]
True

>>> existOndaAtiva [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) 5,Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]
True

>>> existOndaAtiva [Onda [] 5 (-5) (-5)]
False

posicaoOndaAtiva

Arguments

:: [Onda]
Lista de ondas (lista que possui pelo menos uma onda ativa).

-> Int
Posição da onda ativa na lista.

A função posicaoOndaAtiva determina a posição da onda ativa (assumindo que esta existe) numa lista de ondas.

Exemplos

>>> posicaoOndaAtiva [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]
0

>>> posicaoOndaAtiva [Onda [] 40 (-5) (-5),Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]
1

inimigosNoAlcance

Arguments

:: Torre
Torre de foco.

-> [Inimigo]
Lista de inimigos ativos no jogo.

-> [Inimigo]
Lista de inimigos ativos que estão no alcance da torre.

A função inimigosNoAlcance determina todos os inimigos (ativos) que estão no alcance de uma torre.

Exemplos

>>> inimigosNoAlcance (Torre (5,5) 5 2 5 5 5 (Projetil Fraqueza Infinita) 1 1) [Inimigo (2,3) Norte 40 40 40 40 [] [] 1 1,Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1]
[Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1]

>>> inimigosNoAlcance (Torre (5,5) 5 2 5 5 5 (Projetil Fraqueza Infinita) 1 1) []
[]

atingeInimigo

Arguments

:: Torre
Torre de foco.

-> Inimigo
Inimigo atingido.

-> Inimigo
Inimigo resultante.

A função atingeInimigo calcula o resultado de um inimigo ser atingido por uma torre.

Exemplos

>>> atingeInimigo (Torre (5,5) 5 2 5 5 5 (Projetil Fogo (Finita 4)) 1 1) (Inimigo (2,3) Norte 40 40 40 40 [Projetil Gelo Infinita] [Fogo] 1 1)
Inimigo (2,3) Norte 35 40 40 40 [Projetil Gelo Infinita] [Fogo] 1 1

>>> atingeInimigo (Torre (5,5) 5 2 5 5 5 (Projetil Fogo (Finita 4)) 1 1) (Inimigo (2,3) Norte 40 40 40 40 [Projetil Gelo Infinita] [] 1 1)
Inimigo (2,3) Norte 35 40 40 40 [] [] 1 1

>>> atingeInimigo (Torre (5,5) 5 2 5 5 5 (Projetil Fogo (Finita 4)) 1 1) (Inimigo (2,3) Norte 40 40 40 40 [Projetil Fraqueza Infinita] [] 1 1)
Inimigo (2,3) Norte 30 40 40 40 [Projetil Fraqueza Infinita,Projetil Fogo (Finita 4)] [] 1 1

ativaInimigo

Arguments

:: Portal
Portal de foco.

-> [Inimigo]
Lista de inimigos ativos.

-> (Portal, [Inimigo])
Resultado na forma '(portal resultante,lista de inimigos ativos resultante)'.

A função ativaInimigo calcula o que acontece a um portal quando uma das suas ondas ativa um inimigo.

Exemplos

>>> ativaInimigo (Portal (4,4) [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]) []
(Portal (4,4) [Onda [] 5 5 (-5)],[Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1])

>>> ativaInimigo (Portal (4,4) [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) 5,Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]) []
(Portal (4,4) [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 (-5) 5,Onda [] 5 5 (-5)],[Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1])

terminouJogo

Arguments

:: Jogo
Jogo atual.

-> Bool
Resultado na forma True se o jogo já acabou ou False caso contrário.

A função terminouJogo verifica se o jogo terminou, isto é, se o jogador já perdeu ou já ganhou.

Exemplos

>>> terminouJogo (Jogo (Base 50 (4,4) 50) [Portal (4,4) [Onda [] 5 5 5]] [] [] [] [] 1 (1,1) 0 (False,30,[]))
True

>>> terminouJogo (Jogo (Base (-10) (4,4) 50) [] [] [] [] [] 1 (1,1) 0 (False,30,[]))
True

>>> terminouJogo (Jogo (Base 50 (4,4) 50) [Portal (4,4) [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 5 5]] [] [] [] [] 1 (1,1) 0 (False,30,[]))
False

ganhouJogo

Arguments

:: Jogo
Jogo atual.

-> Bool
Resultado na forma True se o jogador já tiver ganho ou False caso contrário.

A função ganhouJogo verifica se o jogador ganhou o jogo.

Exemplos

>>> ganhouJogo (Jogo (Base 50 (4,4) 50) [Portal (4,4) [Onda [] 5 5 5]] [] [] [] [] 1 (1,1) 0 (False,30,[]))
True

>>> ganhouJogo (Jogo (Base 50 (4,4) 50) [Portal (4,4) [Onda [Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1] 5 5 5]] [] [] [] [] 1 (1,1) 0 (False,30,[]))
False

perdeuJogo

Arguments

:: Jogo
Jogo atual.

-> Bool
Resultado na forma True se o jogador já tiver perdido ou False caso contrário.

A função perdeuJogo verifiica se o jogador perdeu o jogo.

Exemplos

>>> perdeuJogo (Jogo (Base 50 (4,4) 50) [] [] [] [] [] 1 (1,1) 0 (False,30,[]))
False

>>> perdeuJogo (Jogo (Base (-10) (4,4) 50) [] [] [] [] [] 1 (1,1) 0 (False,30,[]))
True