Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2024/25.

Documentation

inimigosNoAlcance :: Torre -> [Inimigo] -> [Inimigo]

A funcao inimigosNoAlcance tem como objetivo relacionar a posicao da torre e o seu alcance para definir se um inimigo qualquer está no alcance dessa torre.

Exemplos de Aplicação:

Expand

>>> inimigosNoAlcance (Torre {posicaoTorre=(2,4), alcanceTorre= 10.2, danoTorre=2.0, rajadaTorre=2, cicloTorre=1.0, tempoTorre=1.0, projetilTorre= Projetil Fogo (Finita 3)}) [Inimigo Normal (3,6) Sul 10.0 1.0 5.0 10 [] Walk 1 0 [] Nothing]
[Inimigo {tipoInimigo = Normal, posicaoInimigo = (3.0,6.0), direcaoInimigo = Sul, vidaInimigo = 10.0, velocidadeInimigo = 1.0, ataqueInimigo = 5.0, butimInimigo = 10, projeteisInimigo = [], animacao = Walk, frameInimigo = 1, tempoAInimigo = 0.0, caminhoInimigo = [], posMedia = Nothing}]

>>> inimigosNoAlcance (Torre {posicaoTorre = (2, 4), alcanceTorre = 10.2, danoTorre=2.0, rajadaTorre=2, cicloTorre=1.0, tempoTorre=1.0, projetilTorre= Projetil Fogo (Finita 3)}) [Inimigo Normal (20,20) Sul 10.0 1.0 5.0 10 [] Walk 1 0 [] Nothing]
[]

Propriedades:

Expand

Para cada inimigo na lista de output(dentro do range da torre), a distancia entre sua posiçao e a posicao da torre deve ser menor ou igual ao alcance da torre: all (Inimigo {posicaoInimigo = (xi, yi)} -> sqrt ((xi - xt)^2 + (yi - yt)^2) <= alcance) (inimigosNoAlcance (Torre {posicaoTorre = (xt, yt), alcanceTorre = alcance}) inimigos)

Para cada inimigo na lista de output(fora do range da torre), a distancia entre a sua posicao e a posicao da torre precisa ser maior que o alcance da torre,logo ele nao estara no alcance(range) da torre e, portanto, fora da lista de output: all (Inimigo {posicaoInimigo = (xi, yi)} -> sqrt ((xi - xt)^2 + (yi - yt)^2) > alcance) (inimigos \ inimigosNoAlcance (Torre {posicaoTorre = (xt, yt), alcanceTorre = alcance}) inimigos)

Função Auxiliar:

Expand

A funcao "estaNoAlcance" simplifica o calculo da distancia entre dois pontos(posicaoInimigo e posicaoTorre) para ser utilizada pelo filter

estaNoAlcance Inimigo {posicaoInimigo=(xi, yi)} = 
    sqrt ((xi-xt)^(2::Int)+(yi-yt)^(2::Int))  <= alcance

atingeInimigo :: Torre -> Inimigo -> Inimigo

A funcao atingeInimigo aplica o dano de uma torre a um inimigo específico, reduzindo sua vida com base no dano da torre e atualizando a lista de projéteis ativos no inimigo com o respectivo projétil da torre.

Exemplos de Aplicação:

Expand

>>> atingeInimigo (Torre {danoTorre = 5, projetilTorre = Projetil {tipoProjetil = Fogo, duracaoProjetil=(Finita 3)}, posicaoTorre = (0, 0), alcanceTorre = 10, rajadaTorre=2, cicloTorre=2.0, tempoTorre=2.0}) (Inimigo {tipoInimigo = Normal, posicaoInimigo = (3.0,6.0), direcaoInimigo = Sul, vidaInimigo = 10.0, velocidadeInimigo = 1.0, ataqueInimigo = 5.0, butimInimigo = 10, projeteisInimigo = [], animacao = Walk, frameInimigo = 1, tempoAInimigo = 0.0, caminhoInimigo = [], posMedia = Nothing})
Inimigo {tipoInimigo = Normal, posicaoInimigo = (3.0,6.0), direcaoInimigo = Sul, vidaInimigo = 5.0, velocidadeInimigo = 1.0, ataqueInimigo = 5.0, butimInimigo = 10, projeteisInimigo = [Projetil {tipoProjetil = Fogo, duracaoProjetil = Finita 3.0}], animacao = Walk, frameInimigo = 1, tempoAInimigo = 0.0, caminhoInimigo = [], posMedia = Nothing}

>>> atingeInimigo (Torre {danoTorre = 15, projetilTorre = Projetil {tipoProjetil = Gelo, duracaoProjetil=(Finita 3)}, rajadaTorre=1, cicloTorre=2.0, tempoTorre=1.0, posicaoTorre = (0, 0), alcanceTorre = 10}) (Inimigo {tipoInimigo = Normal, posicaoInimigo = (3.0,6.0), direcaoInimigo = Sul, vidaInimigo = 10.0, velocidadeInimigo = 1.0, ataqueInimigo = 5.0, butimInimigo = 10, projeteisInimigo = [], animacao = Walk, frameInimigo = 1, tempoAInimigo = 0.0, caminhoInimigo = [], posMedia = Nothing})
Inimigo {tipoInimigo = Normal, posicaoInimigo = (3.0,6.0), direcaoInimigo = Sul, vidaInimigo = 0.0, velocidadeInimigo = 1.0, ataqueInimigo = 5.0, butimInimigo = 10, projeteisInimigo = [Projetil {tipoProjetil = Gelo, duracaoProjetil = Finita 3.0}], animacao = Walk, frameInimigo = 1, tempoAInimigo = 0.0, caminhoInimigo = [], posMedia = Nothing}

Propriedades:

Expand

A vida atualizada do inimigo nunca será negativa: novaVida = max 0 (vida - dano)

Funcoes Auxiliares:

Expand

Função atualizaProjeteis

Função combinaComExistentes

atualizaProjeteis :: Projetil -> [Projetil] -> [Projetil]

A função atualizaProjeteis adiciona um novo projétil a uma lista de projéteis ativos de um inimigo.

Exemplos de Aplicação:

Expand

>>> atualizaProjeteis (Projetil {tipoProjetil = Fogo, duracaoProjetil=(Finita 3)}) [Projetil {tipoProjetil = Gelo, duracaoProjetil=(Finita 3)}]
[]
>>> atualizaProjeteis (Projetil {tipoProjetil = Resina, duracaoProjetil=(Finita 3)}) []
[Projetil {tipoProjetil = Resina, duracaoProjetil = Finita 3.0}]

combinaComExistentes :: Projetil -> [Projetil] -> [Projetil]

A função combinaComExistentes verifica se dois projéteis possuem o mesmo tipo e as suas possiveis Sinergias.

Exemplos de Aplicação:

Expand

>>> combinaComExistentes (Projetil {tipoProjetil = Fogo, duracaoProjetil = Finita 3}) [Projetil {tipoProjetil = Fogo, duracaoProjetil = Finita 2}]
[Projetil {tipoProjetil = Fogo, duracaoProjetil = Finita 5.0}]

>>> combinaComExistentes (Projetil {tipoProjetil = Gelo, duracaoProjetil = Infinita}) [Projetil {tipoProjetil = Fogo, duracaoProjetil = Finita 3}]
[Projetil {tipoProjetil = Fogo, duracaoProjetil = Finita 3.0},Projetil {tipoProjetil = Gelo, duracaoProjetil = Infinita}]

ativaInimigo :: Portal -> [Inimigo] -> (Portal, [Inimigo])

A função ativaInimigo tem como objetivo receber um portal, ativando inimigos da onda atual para adicioná-los à lista de inimigos ativos. Se a onda atual do portal não possuir mais inimigos, a função avança para a próxima onda. Caso o portal não tenha ondas restantes, nenhum inimigo é ativado!!

Propriedades:

Expand

Para um portal sem ondas, a lista de inimigos ativos permanece inalterada

Após a ativação, o portal nunca mantém a mesma lista de inimigos na onda atual, pois esses inimigos sao inseridos no mapa

terminouJogo :: Jogo -> Bool

A função terminouJogo verifica se o jogo atingiu uma condição de término. O jogo termina se:

O jogador perdeu, condição indicada pela função auxiliar perdeuJogo.

O jogador ganhou, condição indicada pela função auxiliar ganhouJogo.

Propriedades:

Expand

Um jogo em que nem perdeuJogo nem ganhouJogo é verdadeiro deve retornar False

Um jogo em que terminouJogo é True implica que pelo menos uma das condições de término foi atingida, ou ganhou ou perdeu

ganhouJogo :: Jogo -> Bool

A função ganhouJogo verifica se o jogador venceu o jogo. O jogo é considerado ganho se:

A base do jogador tem vida superior a zero (indicado por vidaBase).

Não há inimigos restantes no jogo (indicados por inimigosJogo do jogo).

Todas as ondas dos portais estão sem inimigos (verificado pela função auxiliar semInimigos).

Propriedades:

Expand

Um jogo onde a base tem vida > 0, não há inimigos e todas as ondas estão sem inimigos é considerado "ganho".

Funcao Auxiliar:

Expand

Função semInimigos

semInimigos :: Portal -> Bool

A função semInimigos verifica se um portal não possui inimigos em suas ondas. Um portal é considerado sem inimigos se todas as ondas dentro do portal não tiverem inimigos nelas.

perdeuJogo :: Jogo -> Bool

A função perdeuJogo verifica se o jogador perdeu o jogo. O jogo é considerado "perdido" se a vida da base do jogador for menor ou igual a zero.

Propriedades:

Expand

Um jogo é perdido quando a vida da base for menor ou igual a zero.