Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2024/25.

Documentation

inimigosNoAlcance :: Torre -> [Inimigo] -> [Inimigo]

Calcula os inimigos ao alcance de uma dada torre.

Exemplos:

Expand

>>> inimigosNoAlcance (Torre (2,1) 6.5 3 1 2 0 (Projetil Fogo (Finita 5)) [Inimigo (1,1) Este 10 1 1 15 [] 1]
[Inimigo (1,1) Este 10 1 1 15 [] 1]
>>> inimigosNoAlcance (Torre (2,1) 6.5 3 1 2 0 (Projetil Fogo (Finita 5)) [Inimigo (1,1) Este 10 1 1 15 [] 1,Inimigo (6,6) Este 10 1 1 15 [] 1]
[Inimigo (1,1) Este 10 1 1 15 [] 1]

atingeInimigo :: Torre -> Inimigo -> Inimigo

Função que atualiza o estado de um inimigo assumindo que este acaba de ser atingido por um projétil de uma torre.

Exemplos:

Expand

>>> atingeInimigo (Inimigo (1,1) Este 10 1 1 15 [] 1) (Torre (2,1) 6.5 3 1 2 0 (Projetil Fogo (Finita 5))
Inimigo (1,1) Este 3.5 1 1 15 [Projetil Fogo (Finita 5)] 1
>>> atingeInimigo (Inimigo (1,1) Este 10 1 1 15 [Projetil Fogo (Finita 5)] 1) (Torre (2,1) 6.5 3 1 2 0 (Projetil Gelo (Finita 5))
Inimigo (1,1) Este 3.5 1 1 15 [] 1

Propriedades:

Expand

Fogo e Gelo cancelam-se mutuamente.

Fogo e Resina dobram a duração do projétil de Fogo.

Projéteis iguais somam as suas durações.

ativaInimigo :: Portal -> [Inimigo] -> (Portal, [Inimigo])

Dado um Portal e a lista de inimigos atualmente em jogo, move o próximo inimigo a ser lançado pelo portal para a lista de inimigos ativos.

Exemplos:

Expand

>>> ativaInimigo (Portal (0,0) [Onda [Inimigo (0,0) Este 10 1 1 15 [] 1] 2 0 0]) []
(Portal (0,0) [Onda [] 2 2 0],[Inimigo (0,0) Este 10 1 1 15 [] 1])
>>> ativaInimigo (Portal (0,0) [Onda [Inimigo (0,0) Este 10 1 1 15 [] 1] 2 2 0]) []
(Portal (0,0) [Onda [Inimigo (0,0) Este 10 1 1 15 [] 1] 2 2 0],[])

terminouJogo :: Jogo -> Bool

Verifica se o jogo terminou (se o jogador ganhou ou perdeu).

Exemplos:

Expand

>>> terminouJogo (Jogo (Base 10 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [] [])
True
>>> terminouJogo Jogo (Base 10 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [Inimigo (1,1) Este 10 1 1 15 [] 1] []
False 
>>> terminouJogo Jogo (Base 0 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [Inimigo (1,1) Este 10 1 1 15 [] 1] []
True 

ganhouJogo :: Jogo -> Bool

Verifica se o jogador ganhou o jogo.

Exemplos:

Expand

>>> ganhouJogo (Jogo (Base 10 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [] [])
True 
>>> ganhouJogo Jogo (Base 10 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [Inimigo (1,1) Este 10 1 1 15 [] 1] []
False 

perdeuJogo :: Jogo -> Bool

Verifica se o jogador perdeu o jogo.

Exemplos:

Expand

>>> perdeuuJogo (Jogo (Base 10 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [] [])
False
>>> perdeuJogo Jogo (Base 0 (0,0) 0) [Portal (1,2) [Onda [] 0 0 0]] [] mapa01 [Inimigo (1,1) Este 10 1 1 15 [] 1] []
True 

multiplicarDuracao :: Duracao -> Int -> Duracao

Multiplica a duração por um numero inteiro.

somarDuracoes :: Duracao -> Duracao -> Duracao

Soma duas durações.

existeGelo :: [Projetil] -> Bool

Verifica se existe um projetil de gelo numa lista de projeteis.

existeResina :: [Projetil] -> Bool

Verifica se existe um projetil de resina numa lista de projeteis.

existeFogo :: [Projetil] -> Bool

Verifica se existe um projetil de fogo numa lista de projeteis.

removerGelo :: [Projetil] -> [Projetil]

Remove um projetil de gelo da lista de projeteis.

removerFogo :: [Projetil] -> [Projetil]

Remove um projetil de fogo da lista de projeteis.

removerResina :: [Projetil] -> [Projetil]

Remove um projetil de resina da lista de projeteis.

fogoProjetil :: [Projetil] -> Projetil

Procura um projetil de fogo numa lista em que exista um projetil de fogo.

geloProjetil :: [Projetil] -> Projetil

Procura um projetil de gelo numa lista em que exista um projetil de fogo.

resinaProjetil :: [Projetil] -> Projetil

Procura um projetil de resina numa lista em que exista um projetil de fogo.

portaisSemInimigos :: [Portal] -> Bool

Verifica se os portais estao sem inimigos.

ondasSemInimigos :: [Onda] -> Bool

Verifica se as ondas estao sem inimigos.

distancia :: Posicao -> Posicao -> Float

Calcula a distancia entre duas posições.