Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2024/25.

Documentation

atualizaJogo

Arguments

:: Tempo
Tempo a atualizar.

-> Tempo
Tempo de jogo.

-> Jogo
Jogo pré-atualização.

-> Jogo
Jogo atualizado.

A função atualizaJogo é responsável por atualizar tudo o que ocorre durante um jogo automaticamente, tal como o movimentos dos inimigos e os ataques das torres.

mapa01 :: Mapa

mapa01 é um exemplo de um mapa.

mapaX :: Mapa

mapaX é um exemplo de um mapa.

mapaY :: Mapa

mapaY é um exemplo de um mapa.

atualizaEfeitosChuva

Arguments

:: [Inimigo]
Inimigos pré-atualização.

-> Int
Número do nível do jogo.

-> Bool
Se a chuva está ou não ativa.

-> Tempo
Tempo a atualizar.

-> [Inimigo]
Inimigos pós-atualização.

A função atualizaEfeitosChuva atualiza os inimigos com base nos efeitos da chuva.

Exemplos

>>> atualizaEfeitosChuva [Inimigo (4,4) Norte 40 40 40 40 [Projetil Fogo Infinita] [] 1 1] 1 False 5
[Inimigo (4,4) Norte 40 40 40 40 [Projetil Fogo Infinita] [] 1 1]

>>> atualizaEfeitosChuva [Inimigo (4,4) Norte 40 40 40 40 [Projetil Fogo Infinita] [] 1 1] 1 True 5
[Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1]

>>> atualizaEfeitosChuva [Inimigo (4,4) Norte 40 40 40 40 [Projetil Fogo Infinita] [] 1 1] 2 True 5
[Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1]

>>> atualizaEfeitosChuva [Inimigo (4,4) Norte 40 40 40 40 [] [Fogo] 1 1,Inimigo (4,4) Norte 40 40 40 40 [Projetil Gelo Infinita] [] 1 1,Inimigo (4,4) Norte 40 40 40 40 [Projetil Resina Infinita] [] 1 1] 3 True 5
[Inimigo (4,4) Norte 40 40 40 40 [] [Fogo] 1 1,Inimigo (4,4) Norte 40 40 40 40 [Projetil Fogo (Finita 5)] [] 1 1,Inimigo (4,4) Norte 40 40 40 40 [Projetil Fogo (Finita 5)] [] 1 1]

>>> atualizaEfeitosChuva [Inimigo (4,4) Norte 30 40 40 40 [] [] 100 1,Inimigo (4,4) Norte 90 40 40 40 [] [] 100 1] 4 True 5
[Inimigo (4,4) Norte 80 40 40 40 [] [] 100 1,Inimigo (4,4) Norte 100 40 40 40 [] [] 100 1]

regeneraInimigoChuvaLagrimasCristalinas :: Tempo -> Inimigo -> Inimigo

A função regeneraInimigoChuvaLagrimasCristalinas regenera os inimigos quando está a chover lágrimas cristalinas.

Exemplos

>>> regeneraInimigoChuvaLagrimasCristalinas 5 (Inimigo (4,4) Norte 30 40 40 40 [] [] 100 1)
Inimigo (4,4) Norte 80 40 40 40 [] [] 100 1

>>> regeneraInimigoChuvaLagrimasCristalinas 5 (Inimigo (4,4) Norte 90 40 40 40 [] [] 100 1)
Inimigo (4,4) Norte 100 40 40 40 [] [] 100 1

>>> regeneraInimigoChuvaLagrimasCristalinas 5 (Inimigo (4,4) Norte 100 40 40 40 [] [] 100 1)
Inimigo (4,4) Norte 100 40 40 40 [] [] 100 1

filtraPosicoesLagrimasCristalinas

Arguments

:: Mapa
Mapa do jogo.

-> Posicao
Posição do ponto na matriz (deve ser (0,0) inicialmente para determinar corretamente as coordenadas).

-> [Posicao]
Posições correpondentes ao terreno de LagrimasCristalinas.

A função filtraPosicoesLagrimasCristalinas determina as coordenadas de pontos correnpondentes apenas ao terreno de LagrimasCristalinas de um mapa.

Exemplos

>>> filtraPosicoesLagrimasCristalinas mapaX (0,0)
[]

>>> filtraPosicoesLagrimasCristalinas mapaY (0,0)
[(2,0),(1,1),(2,1)]

verifGelo

Arguments

:: Projetil
Projétil recebido.

-> Bool
Resultado no tipo True caso o projétil seja de Gelo ou False caso contrário.

A função verifGelo verifica se um projétil é do tipo Gelo.

Exemplos

>>> verifGelo (Projetil Gelo Infinita)
True

>>> verifFogo (Projetil Resina Infinita)
False

verifFogo

Arguments

:: Projetil
Projétil recebido.

-> Bool
Resultado no tipo True caso o projétil seja de Fogo ou False caso contrário.

A função verifFogo verifica se um projétil é do tipo Fogo.

Exemplos

>>> verifFogo (Projetil Fogo Infinita)
True

>>> verifFogo (Projetil Resina Infinita)
True

verifResina

Arguments

:: Projetil
Projétil recebido.

-> Bool
Resultado no tipo True caso o projétil seja de Resina ou False caso contrário.

A função verifResina verifica se um projétil é do tipo Resina.

Exemplos

>>> verifResina (Projetil Resina Infinita)
True

>>> verifResina (Projetil Fraqueza Infinita)
False

inimigoNaoEstaNoAlcance

Arguments

:: Torre
Torre de escolha.

-> Inimigo
Inimigo de escoha.

-> Bool
Resultado na forma True caso o inimigo não esteja no alcance da torre ou False caso contrário.

A função inimigoNaoEstaNoAlcance verifica se um inimigo não está no alcance de uma torre.

Exemplos

>>> inimigoNaoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (5,4) Norte 50 50 50 50 [] [] 50 1)
False

>>> inimigoNaoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1)
False

>>> inimigoNaoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (9,4) Norte 50 50 50 50 [] [] 50 1)
True

>>> inimigoNaoEstaNoAlcance (Torre (4,4) 5 4 5 5 5 (Projetil Fogo Infinita) 1 1) (Inimigo (8,8) Norte 50 50 50 50 [] [] 50 1)
True

detetaInimigos

Arguments

:: Torre
Torre a atualizar.

-> [Inimigo]
Inimigos ativos que se encontram no alcance da torre.

-> Tempo
Tempo a atualizar.

-> (Torre, [Inimigo])
Resultado com a torre e os inimigos atualizados.

A função detetaInimigos realiza os ataques de uma torre aos inimigos que estão no seu alcance (quantidade máxima correspondente à sua rajada) caso esta esteja pronta para atacar.

Exemplo

>>> detetaInimigos (Torre (4,4) 5 4 2 5 (-5) (Projetil Fogo Infinita) 1 1) [Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1] 5
(Torre (4,4) 5 4 2 5 5 (Projetil Fogo Infinita) 1 1,[Inimigo (5,5) Norte 45 50 50 50 [Projetil Fogo Infinita] [] 50 1,Inimigo (5,5) Norte 45 50 50 50 [Projetil Fogo Infinita] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1])]

removeInimigosMortos

Arguments

:: [Inimigo]
Lista de inimigos.

-> Base
Base do jogo.

-> Int
Quantidade de inimigos derrotados.

-> (Base, [Inimigo], Int)
Resultado com a base atualizada e os inimigos com vida positiva (vivos).

A função removeInimigosMortos remove os inimigos mortos do jogo e, como consequência, altera os créditos da base.

Exemplo

>>> removeInimigosMortos [Inimigo (5,5) Norte (-10) 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 0 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1] (Base 50 (4,4) 50) 10
(Base 100 (4,4) 100,[Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1])]

inimigosAtingemBase

Arguments

:: Base
Base do jogo.

-> [Inimigo]
Lista de inimigos.

-> (Base, [Inimigo])
Resultado com a base atualizada e os inimigos que ainda não atingiram a base.

A função inimigosAtingemBase remove os inimigos que já alcançaram a base e danifica a mesma como consequência.

Exemplo

>>> [inimigosAtingemBase (Base 200 (4,4) 100) [Inimigo (4.25,4) Norte 50 50 50 50 [] [] 50 1,Inimigo (4.25,4) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1]
(Base 100 (4,4) 100,[Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1,Inimigo (5,5) Norte 50 50 50 50 [] [] 50 1])]

atualizaPosicao

Arguments

:: Direcao
Direção do inimigo.

-> Posicao
Posição do inimigo.

-> Tempo
Tempo a atualizar.

-> Float
Velocidade do inimigo (afetada pelos projéteis).

-> Posicao
Posição do inimigo atualizada.

A função atualizaPosicao atualiza a posição de um inimigo.

Exemplos

>>> atualizaPosicao Norte (4,4) 3 1
(4,1)

>>> atualizaPosicao Sul (4,4) 3 1
(4,7)

>>> atualizaPosicao Este (4,4) 3 1
(7,4)

>>> atualizaPosicao Oeste (4,4) 3 1
(1,4)

divFloat

Arguments

:: Float
Número positivo.

-> Float
Parte decimal do número.

A função divFloat dá a parte decimal de um número positivo.

Exemplo

>>> divFloat 4.5
0.5

atualizaDirecao

Arguments

:: Direcao
Direção do inimigo.

-> Posicao
Posição do inimigo.

-> Posicao
Posição da base.

-> [Posicao]
Lista de posições dos caminhos de terra do mapa.

-> Direcao
Direção do inimigo atualizada.

A função atualizaDirecao atualiza a direção de um inimigo de forma a que não o permita sair dos caminhos de terra (a direção do inimigo é atualizada apenas quando este se encontra na região central de uma terra, pelo que todos os inimigos devem começar no centro de uma terra).

Exemplos

>>> atualizaDirecao Norte (0.5,0.5) (5,2) [(0.0,0.0),(1.0,0.0),(4.0,0.0),(1.0,1.0),(4.0,1.0),(1.0,2.0),(5.0,2.0),(1.0,3.0),(2.0,3.0),(5.0,3.0),(1.0,4.0),(2.0,4.0),(3.0,4.0),(4.0,4.0),(5.0,4.0),(1.0,5.0)]
Este

>>> atualizaDirecao Este (1.5,0.5) (5,2) [(0.0,0.0),(1.0,0.0),(4.0,0.0),(1.0,1.0),(4.0,1.0),(1.0,2.0),(5.0,2.0),(1.0,3.0),(2.0,3.0),(5.0,3.0),(1.0,4.0),(2.0,4.0),(3.0,4.0),(4.0,4.0),(5.0,4.0),(1.0,5.0)]
Sul

>>> atualizaDirecao Sul (1.5,4.5) (5,2) [(0.0,0.0),(1.0,0.0),(4.0,0.0),(1.0,1.0),(4.0,1.0),(1.0,2.0),(5.0,2.0),(1.0,3.0),(2.0,3.0),(5.0,3.0),(1.0,4.0),(2.0,4.0),(3.0,4.0),(4.0,4.0),(5.0,4.0),(1.0,5.0)]
Este

atualizaInimigo

Arguments

:: Tempo
Tempo a atualizar.

-> Mapa
Mapa do jogo.

-> Posicao
Posição da base.

-> Bool
Se está ou não a chover.

-> Inimigo
Inimigo a atualizar.

-> Inimigo
Inimigo atualizado.

A função atualizaInimigo atualiza a posição, a vida e a direção de um inimigo com base nas suas circuntâncias (projéteis, posição no mapa, entre outros fatores).

Exemplos

>>> atualizaInimigo 1 mapa01 (5,2) False (Inimigo (0.5,0.5) Norte 100 1 10 10 [Projetil Fogo (Finita 3)] [] 100 1)
Inimigo (1.5,0.5) Este 90 1 10 10 [Projetil Fogo (Finita 2)] [] 100 1

>>> atualizaInimigo 1 mapa01 (5,2) False (Inimigo (0.5,0.5) Norte 100 1 10 10 [Projetil Gelo (Finita 3)] [] 100 1)
Inimigo (0.5,0.5) Norte 100 1 10 10 [Projetil Gelo (Finita 2)] [] 100 1

>>> atualizaInimigo 1 mapa01 (5,2) False (Inimigo (0.5,0.5) Norte 100 1 10 10 [Projetil Resina Infinita] [] 100 1)
Inimigo (1.2,0.5) Este 100 1 10 10 [Projetil Resina Infinita] [] 100 1

atualizaProjetil

Arguments

:: Tempo
Tempo a atualizar.

-> [Projetil]
Lista de projéteis a atualizar.

-> [Projetil]
Lista de projéteis atualizados.

A função atualizaProjetil atualiza a duração de um projétil (caso seja duração finita).

Exemplos

>>> atualizaProjetil 1 [Projetil Fogo Infinita,Projetil Gelo (Finita 0.5)]
[Projetil Fogo Infinita]

>>> atualizaProjetil 1 [Projetil Fogo Infinita,Projetil Gelo (Finita 1.5)]
[Projetil Fogo Infinita,Projetil Gelo (Finita 0.5)]

atualizaPortal

Arguments

:: Portal
Portal a atualizar.

-> Tempo
Tempo a atualizar.

-> [Inimigo]
Lista de inimigos ativos do jogo.

-> (Portal, [Inimigo])
Resultado com o portal atualizado e todos os inimigos ativos do jogo.

A função atualizaPortal atualiza um portal de modo a que coloque em jogo inimigos das suas ondas quando estiver pronto.

Exemplos

>>> atualizaPortal (Portal (4,4) [Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 (-5) 5]) 1 []
(Portal (4,4) [Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 (-6) 4],[])

>>> atualizaPortal (Portal (4,4) [Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]) 1 []
atualizaPortal (Portal (4,4) [Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 (-5) (-5)]) 1 []

mudaPosicaoInimigo :: Inimigo -> Posicao -> Inimigo

A função mudaPosicaoInimigo muda a posição de um inimigo.

Exemplo

>>> mudaPosicaoInimigo (Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1) (4,4)
Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1

atualizaOndaAtiva

Arguments

:: Onda
Onda (ativa) a atualizar.

-> Tempo
Tempo a atualizar.

-> Posicao
Posição do portal.

-> (Onda, [Inimigo])
Resultado com a onda atualizada e a lista de inimigos resultante.

A função atualizaOndaAtiva atualiza uma onda ativa, colocando em jogo um inimigo se estiver pronta.

Exemplos

>>> atualizaOndaAtiva (Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 5 (-5)) 1 (4,4)
(Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 4 (-5),[])

>>> atualizaOndaAtiva (Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 (-4) (-5)) 1 (4,4)
(Onda [] 5 5 (-5),[Inimigo (4,4) Norte 40 40 40 40 [] [] 1 1])

atualizaTemposOnda

Arguments

:: Tempo
Tempo a atualizar.

-> Onda
Onda (não ativa) a atualizar.

-> Onda
Onda atualizada.

A função atualizaTemposOnda atualiza os tempos de uma onda (não ativa).

Exemplos

>>> atualizaTemposOnda 1 (Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 5 100)
Onda [Inimigo (1,1) Norte 40 40 40 40 [] [] 1 1] 5 4 99

atualizaVidaInimigosLagrimasCristalinas

Arguments

:: [Posicao]
Posições das águas cristalinas.

-> [Inimigo]
Inimigos ativos do jogo.

-> Tempo
Tempo a atualizar.

-> (Bool, Tempo, [Posicao])
Tempo de jogo.

-> [Inimigo]
Inimigos eventualmente regenerados.

A função atualizaVidaInimigosLagrimasCristalinas aumenta a vida dos inimigos (ativos) próximos de lágrimas cristalinas.

Exemplos

>>> atualizaVidaInimigosLagrimasCristalinas [(1,1)] [Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1] 1 (True,30,[])
[Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1]

>>> atualizaVidaInimigosLagrimasCristalinas [(1,1)] [Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1] 1 (False,30,[])
[Inimigo (1,1) Norte 45 40 40 40 [] [] 45 1]

>>> atualizaVidaInimigosLagrimasCristalinas [(5,5)] [Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1] 1 (True,30,[])
[Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1]

atualizaVidaInimigoLagrimasCristalinas

Arguments

:: [Posicao]
Posicões das águas cristalinas.

-> Tempo
Tempo a atualizar.

-> Inimigo
Inimigo selecionado.

-> Inimigo
Inimigo eventualmente regenerado.

A função atualizaVidaInimigoLagrimasCristalinas aumenta a vida de um inimigo caso este esteja próximo de lágrimas cristalinas.

Exemplos

>>> atualizaVidaInimigoLagrimasCristalinas [(1,1)] 1 (Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1)
Inimigo (1,1) Norte 45 40 40 40 [] [] 45 1

>>> atualizaVidaInimigoLagrimasCristalinas [(5,5)] 1 (Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1)
Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1

removeGotas

Arguments

:: [Posicao]
Posições das gotas.

-> [Posicao]
Lista resultante de remover as gotas que já estão demasiado baixas.

A função removeGotas remove de uma lista de gotas de chuva aquelas que se situam fora do mapa (em relação às ordenadas).

Exemplo

>>> removeGotas [(0,-900),(0,-200)]
[(0,-200)]

atualizaPosicoesGotas

Arguments

:: [Posicao]
Posições das gotas.

-> Tempo
Tempo a atualizar.

-> [Posicao]
Posições atualizadas.

A função atualizaPosicoesGotas atualiza as posições das gotas.

Exemplo

>>> atualizaPosicoesGotas [(100,-300),(200,-400)] 1
[(95,-400),(195,-500)]

removeProjetilFogo

Arguments

:: Inimigo
Inimigo de escolha.

-> Inimigo
Inimigo sem projéteis de fogo.

A função removeProjetilFogo remove os projéteis de fogo de um inimigo.

Exemplo

>>> removeProjetilFogo (Inimigo (1,1) Norte 40 40 40 40 [Projetil Fogo Infinita,Projetil Gelo Infinita,Projetil Resina Infinita,Projetil Fraqueza Infinita] [] 45 1)
Inimigo (1,1) Norte 40 40 40 40 [Projetil Gelo Infinita,Projetil Resina Infinita,Projetil Fraqueza Infinita] [] 45 1

removeProjetilGelo

Arguments

:: Inimigo
Inimigo de escolha.

-> Inimigo
Inimigo sem projéteis de fogo.

A função removeProjetilGelo remove os projéteis de gelo de um inimigo.

Exemplo

>>> removeProjetilGelo (Inimigo (1,1) Norte 40 40 40 40 [Projetil Fogo Infinita,Projetil Gelo Infinita,Projetil Resina Infinita,Projetil Fraqueza Infinita] [] 45 1)
Inimigo (1,1) Norte 40 40 40 40 [Projetil Fogo Infinita,Projetil Resina Infinita,Projetil Fraqueza Infinita] [] 45 1

adicionaProjetilFogo

Arguments

:: Tempo
Tempo a atualizar.

-> Inimigo
Inimigo de escolha.

-> Inimigo
Inimigo com projétil do fogo adicionado.

A função adicionaProjetilFogo adiciona um projétil de fogo a um inimigo.

Exemplos

>>> adicionaProjetilFogo 1 (Inimigo (1,1) Norte 40 40 40 40 [] [] 45 1)
Inimigo (1,1) Norte 40 40 40 40 [Projetil Fogo (Finita 1)] [] 45 1

>>> adicionaProjetilFogo 1 (Inimigo (1,1) Norte 40 40 40 40 [] [Fogo] 45 1)
Inimigo (1,1) Norte 40 40 40 40 [] [Fogo] 45 1

removeProjetilResina

Arguments

:: Inimigo
Inimigo de escolha.

-> Inimigo
Inimigo sem projéteis de resina.

A função removeProjetilResina remove os projéteis de resina de um inimigo.

Exemplo

>>> removeProjetilResina (Inimigo (1,1) Norte 40 40 40 40 [Projetil Fogo Infinita,Projetil Gelo Infinita,Projetil Resina Infinita,Projetil Fraqueza Infinita] [] 45 1)
Inimigo (1,1) Norte 40 40 40 40 [Projetil Fogo Infinita,Projetil Gelo Infinita,Projetil Fraqueza Infinita] [] 45 1

atualizaTorres

Arguments

:: [Torre]
Lista de torres do jogo.

-> [Inimigo]
Lista de inimigos ativos do jogo.

-> Tempo
Tempo a atualizar.

-> ([Torre], [Inimigo])
Resultado com as torres e os inimigos ativos do jogo atualizados.

A função atualizaTorres permite que as torres realizem devidamente os seus ataques aos inimigos ativos do jogo.

Exemplo

>>> atualizaTorres [Torre (1,1) 5 1 1 5 0 (Projetil Fogo Infinita) 1 1,Torre (4,4) 5 1 1 5 1 (Projetil Fogo Infinita) 1 1] [Inimigo (1,1) Norte 40 40 40 40 [] [] 40 1,Inimigo (4,4) Norte 40 40 40 40 [] [] 40 1] 1
([Torre (1,1) 5 1 1 5 5 (Projetil Fogo Infinita) 1 1,Torre (4,4) 5 1 1 5 0 (Projetil Fogo Infinita) 1 1],[Inimigo (4,4) Norte 40 40 40 40 [] [] 40 1,Inimigo (1,1) Norte 35 40 40 40 [Projetil Fogo Infinita] [] 40 1])

atualizaInimigos

Arguments

:: [Inimigo]
Inimigos ativos do jogo.

-> Mapa
Mapa do jogo.

-> Base
Base do jogo.

-> Tempo
Tempo a atualizar.

-> Int
Quantidade de inimigos derrotados.

-> Bool
Se a chuva está ou não ativa.

-> (Base, [Inimigo], Int)
Resultado com a base e os inimigos atualizados.

A função atualizaInimigos atualiza os inimigos do jogo ao remover os mortos e os que já alcançaram a base (assim como as consequências destes acontecimentos) e ao atualizar o movimento dos restantes.

Exemplo

>>> atualizaInimigos [Inimigo (0.5,0.5) Norte 5 1 40 40 [Projetil Fogo Infinita] [] 40 1,Inimigo (0.5,0.5) Norte 40 1 40 40 [] [] 40 1] mapa01 (Base 100 (1,4) 100) 1 5 False
(Base 100 (1,4) 100,[Inimigo (1.5,0.5) Este (-5) 1 40 40 [Projetil Fogo Infinita] [] 40 1,Inimigo (1.5,0.5) Este 40 1 40 40 [] [] 40 1],5)

atualizaPortais

Arguments

:: [Portal]
Lista de portais de jogo.

-> Tempo
Tempo a atualizar.

-> [Inimigo]
Lista de inimigos ativos do jogo.

-> ([Portal], [Inimigo])
Resultado com os portais atualizados e todos os inimigos ativos do jogo.

A função atualizaPortais atualiza os portais de modo a que coloquem em jogo inimigos das suas ondas quando estiverem prontos.

Exemplo

>>> atualizaPortais [Portal (4,4) [Onda [Inimigo (0.5,0.5) Norte 5 1 40 40 [Projetil Fogo Infinita] [] 40 1] 5 0 (-5)]] 1 []
([Portal (4,4) [Onda [] 5 5 (-5)]],[Inimigo (4,4) Norte 5 1 40 40 [Projetil Fogo Infinita] [] 40 1])

atualizaGotasChuva

Arguments

:: (Bool, Tempo, [Posicao])
Estado inicial da chuva.

-> Tempo
Tempo de jogo (multiplicado por 60).

-> Tempo
Tempo a atualizar.

-> (Bool, Tempo, [Posicao])
Estado atualizado da chuva.

A função atualizaGotasChuva atualiza a chuva no jogo.