Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2024/25.

Documentation

atualizaJogo

Arguments

:: Tempo
Recebe o Tempo decorrido do Jogo

-> Jogo
Recebe o próprio Jogo

-> Jogo
Devolve o Jogo com os seus valores atualizados

A função atualizaJogo é responsável por realizar todas as atualizações necessárias do jogo, com o decorrer do tempo. São Integradas diversas funções auxiliares que ajudam no processo, como a movimentação dos inimigos, a ativação dos portais, o ataque das torres e o processo de atualização da base. As atualizações são feitas com base no Tempo decorrido.

_Processo geral:_

Atualiza os portais ativos e gera novos inimigos a partir deles.

Atualiza as torres, considerando os inimigos atingidos e os danos aplicados junto dos seus efeitos.

Atualiza a movimentação dos inimigos no mapa, considerando as modificações dos inimigos após as interações com as torres.

Processa os inimigos restantes, removendo os que foram derrotados e atualizando a base de acordo com os inimigos que ainda estão no jogo.

Atualiza os portais ativos após a passagem de tempo.

Essa função garante que todas as dinâmicas do jogo estejam em constante atualização conforme o tempo avança, refletindo as interações entre inimigos, torres, portais e a base do jogador.

atualizaTorres

Arguments

:: Tempo
Tempo decorrido desde a última atualização das torres

-> [Torre]
Lista de torres presentes no jogo

-> [Inimigo]
Lista de inimigos presentes no jogo

-> ([Torre], [Inimigo])
Retorna a lista de torres atualizadas e a lista de inimigos atualizada.

A função atualizaTorres é responsável por configurar o comportamento das torres em relação aos inimigos em um intervalo de tempo determinado; a cada iteração, ela verifica os inimigos no alcance de cada torre da lista, dispara contra eles e atualiza os estados das torres e dos inimigos com base no tempo do gloss.

Funções Auxiliares:

Expand

inimigosNoAlcance: Filtra os inimigos que estão dentro do alcance de uma torre específica.

disparaTorre: Aplica os danos da torre nos inimigos que estão dentro de seu alcance e atualiza o estado da torre.

Propriedades:

Expand

Caso Base: atualizaTorres _ [] inimigos = ([], inimigos) - Quando não há torres para atualizar, a função retorna uma lista vazia de torres e a lista original de inimigos.

Caso Recursivo: atualizaTorres tempo (torre:restoTorres) inimigos - Filtra os inimigos no alcance da primeira torre usando inimigosNoAlcance. - Atualiza a torre e os inimigos atingidos usando disparaTorre. - Aplica recursivamente a função para o restante das torres e os inimigos atualizados.

disparaTorre

Arguments

:: Tempo
Tempo decorrido desde a última atualização da torre

-> Torre
A torre que está realizando o ataque

-> [Inimigo]
Lista de inimigos presentes no jogo

-> (Torre, [Inimigo])
Retorna a torre atualizada e a lista de inimigos após o ataque

A função disparaTorre é responsável por configurar o estado de uma torre e aplica seus efeitos nos inimigos dentro do seu alcance. Ela considera o tempo decorrido para reduzir o tempo de recarga de UMA torre e aplica dano e efeitos nos inimigos atingidos. Ela, resumidamente, configura as mecânicas de uma torre, para que a atualizaTorres "execute" para todas as torres.

Processo geral:

Expand

Verifica se a torre está em recarga: - Se estiver, apenas atualiza o tempo de recarga restante. - Caso contrário, passa para o próximo passo.

Identifica os inimigos que estão no alcance da torre.

Seleciona até um número máximo de inimigos a serem atingidos, com base no parâmetro rajadaTorre.

Aplica os danos e efeitos da torre aos inimigos selecionados.

Atualiza a lista de inimigos, substituindo os atingidos por suas versões atualizadas.

Reinicia o tempo de recarga da torre após o ataque.

atualizaInimigos

Arguments

:: Mapa
Recebe o Mapa do Jogo, utilizado para validar as novas posições dos inimigos(AKA Terreno)

-> Posicao

-> Tempo
Recebe o tempo decorrido de Jogo

-> [Inimigo]
Recebe a lista de Inimigos em Jogo

-> [Inimigo]
Devolve a lista de Inimigos movimentados

A função "atualizaInimigos" é responsável por configurar a atualização completa em todos os inimigos no mapa, considerando os efeitos ativos de projéteis que podem alterar sua velocidade e vida, e ajustando suas posições conforme a direção e o terreno.

Processo geral:

A função "atualizaInimigos" aplica a função auxiliar "atualizaInimigo" a cada inimigo da lista usando a funçãomap.

"atualizaInimigo" processa um único inimigo, ajustando sua posição e projetando-o da seguinte forma: - Atualiza os projéteis usando atualizaDuracaoProjeteis, - Configura o efeito de fogo usando danoFogo, - Calcula a velocidade ajustada baseada nos efeitos dos projéteis(velocidadeAjustada), - Determina a nova posição considerando a direção e velocidade ajustada, - Verifica se a nova posição é válida no mapa: - Se válida: Atualiza a posição do inimigo, - Se inválida: Altera sua direção com "mudaDirecao".

Funções Auxiliares:

Expand

"atualizaInimigo": Atualiza a posição e os efeitos ativos de UM único inimigo, garantindo que a posição seja válida no mapa.

"atualizaDuracaoProjeteis": Atualiza a duração dos projéteis aplicados ao inimigo, removendo efeitos expirados.

AplicarDanoFogo: Responsável pela configuração do dano por segundo do elemento fogo.

"ajustaVelocidade": Calcula a velocidade ajustada do inimigo, considerando a velocidade base e os efeitos dos projéteis.

"posicaoValida": Verifica se a nova posição do inimigo é válida no mapa (por exemplo, não é um terreno de água).

"mudaDirecao": Altera a direção do inimigo caso a posição calculada seja inválida.

atualizaInimigo

Arguments

:: Mapa
Recebe o mapa do jogo

-> Posicao
Recebe a posição da base no mapa

-> Tempo
Recebe o tempo decorrido

-> Inimigo
Recebe o inimigo a ser atualizado

-> Inimigo
Devolve o inimigo atualizado

A função atualizaInimigo é responsável por calcular o novo estado de um inimigo após a passagem de um intervalo de tempo. Esta atualiza a posição, a vida, os efeitos dos projéteis, e o trajeto do inimigo em direção à base.

Processo geral:

Atualiza os projéteis que estão a afetar o inimigo, reduzindo as suas durações.

Calcula o dano causado pelos projéteis (como efeitos de fogo) e ajusta a vida do inimigo.

Ajusta a velocidade do inimigo com base nos projéteis que o atingem.

Verifica se o inimigo já tem um caminho definido: - Se sim, segue o caminho restante. - Se não, calcula um novo caminho em direção à base, caso possível.

Atualiza a posição do inimigo com base na sua velocidade e no tempo decorrido: - Se o progresso em direção ao próximo ponto do caminho for maior ou igual a 1, passa para o próximo ponto. - Caso contrário, calcula uma posição intermediária entre os pontos.

Atualiza a direção e o estado de animação do inimigo com base no movimento real.

Essa função garante que os inimigos sigam suas rotas no mapa e sejam afetados pelos elementos do jogo, como projéteis e obstáculos.

encontraCaminho

Arguments

:: Mapa
Recebe o mapa do Jogo

-> Posicao
Recebe a posição inicial (inicialmente o portal)

-> Posicao
Recebe a posição final (Base do jogador)

-> Maybe [Posicao]
Devolve, se existir, uma lista de posições a seguir

A função encontraCaminho gera o caminho em Terra mais curto, que ligue um Portal e a Base.

Processo geral:

Expand

Inicia a busca a partir da posição inicial, marcando-a como visitada e registando-a no caminho.

Em cada tile, verifica os vizinhos válidos da posição atual: - Um vizinho é válido se está dentro dos limites do mapa, é acessível (em terra) e se ainda não foi visitado.

Adiciona os vizinhos válidos à lista de posições a serem processadas.

Quando o Portal é atingido, retorna o caminho percorrido até ele.

Se não há caminho possível, retorna Nothing.

Esta função é usada para guiar os inimigos no jogo, permitindo que eles sigam o trajeto até a base.

calcPosIntermedia

Arguments

:: Posicao
Recebe uma posição inicial

-> Posicao
Recebe uma posição final

-> Float
Recebe o progresso entre os dois pontos

-> Posicao
Devolve a posição intermédia

A função calcPosIntermedia calcula uma posição intermediária entre duas posições, baseando-se num parâmetro de progresso t. O valor de t deve estar no intervalo [0,1], onde 0 corresponde à posição inicial e 1 à posição final.

Esta função é necessária para evitar saltos entre as movimentações dos inimigos, fazendo com que seja mais suave.

Exemplo de Aplicação:

Expand

>>> calcPosIntermedia (0, 0) (10, 10) 0.5
(5.0,5.0)

movePosicao

Arguments

:: Posicao
Recebe a posição inicial

-> Posicao
Recebe a posição final

-> Float
Recebe a distáncia máxima de movimento

-> Posicao
Devolve a nova posição após o movimento

A função movePosicao calcula a nova posição de um objeto em movimento, partindo de uma posição inicial em direção a uma posição final, considerando uma distância máxima de movimento.

Caso o objeto esteja muito próximo do destino (menos de 0.01 unidades), este move-se diretamente para o destino.

Caso contrário, a função normaliza o vetor direção e aplica o movimento proporcional à distância.

Exemplo de Aplicação:

Expand

>>> movePosicao (0, 0) (10, 0) 3
(3.0,0.0)

calculaDirecao

Arguments

:: Posicao
Recebe a posição inicial

-> Posicao
Recebe a posição final

-> Direcao
Devolve a nova direção

A função calculaDirecao determina a direção entre duas posições, com base nas coordenadas suas coordenadas.

Processo geral:

Expand

Se a posição final está à direita da inicial, a direção é Este.

Se está à esquerda, a direção é Oeste.

Se está acima, a direção é Norte.

Caso contrário, a direção é Sul.

distancia

Arguments

:: Posicao
Recebe um ponto

-> Posicao
Recebe um segundo ponto

-> Float
Devolve a distância entre eles

A função distancia calcula a distância entre dois pontos no plano.

Exemplo de Aplicação:

Expand

>>> distancia (0, 0) (3, 4)
5.0

atualizaDuracaoProjeteis

Arguments

:: Tempo
Recebe o tempo decorrido desde a última atualização

-> [Projetil]
Recebe a lista de projeteis a ser atualizada

-> [Projetil]
Devolve a lista de projeteis a ser atualizada

A função atualizaDuracaoProjeteis atualiza a duração dos projéteis no jogo com base no tempo decorrido. Projéteis com duração finita sofrei uma redução da sua duração e são removidos se esta atingir 0. Projéteis com duração infinita não são alterados.

Processo geral:

Expand

Reduz a duração de projéteis com duração finita.

Remove projéteis cuja duração chegou a zero.

ajustaVelocidade

Arguments

:: Float
Recebe a velocidade inicial do inimigo

-> [Projetil]
Recebe a lista de projeteis do inimigo

-> Float
Devolve a velocidade ajustada

A função ajustaVelocidade é responsável por calcular a velocidade ajustada com relação aos impactos de uma lista de projéteis. Sempre levando em consideração modificadores específicos de cada projétil!

Configuração:

Expand

**Modificadores de Velocidade:** | modificador = calculaModificador projetil - Para cada projétil, a função calculaModificador é chamada para calcular o impacto específico sobre a velocidade.

**Soma dos Modificadores:** - Os modificadores de velocidade de todos os projéteis são somados para ajustar a velocidade final.

**Velocidade Final:** - A função garante que a velocidade não possa ser negativa, aplicando 'max 0' à soma dos impactos.

calculaModificador

Arguments

:: Float
Recebe a velocidade atual do inimigo

-> Projetil
Recebe um projetil da lista de projeteis do inimigo

-> Float
Devolve a redução de velocidade necessária

A função calculaModificador é responsável por determinar o impacto de um projétil sobre a velocidade, baseado em seu tipo e tempo restante de sua duração.

Configuração:

Expand

**Projétil do Tipo Resina:** - Se o tempo restante (t) for maior que zero, aplica uma redução de 0.6 à velocidade. - Caso contrário, não aplica impacto.

**Projétil do Tipo Gelo:** - Se o tempo restante (t) for maior que zero, aplica uma redução de 1 à velocidade. - Caso contrário, não aplica impacto.

**Outros Tipos de Projéteis:** - Não aplicam nenhum modificador à velocidade, retornando 0.

posicaoValida

Arguments

:: Mapa
Recebe o mapa do jogo

-> Posicao
Recebe uma posicao

-> Bool
Devolve um Booleano

A função posicaoValida é responsável por verificar se uma posição em um mapa é válida, ou seja, se o terreno na posição especificada não é de água ou relva!!

Configuração:

Expand

**Verificação de Terreno:** - A função acessa o mapa na posição correspondente (x, y) e verifica o tipo de terreno. - Se o terreno for Agua ou Relva, a posição é considerada inválida.

**Acesso ao Mapa:** - A função utiliza o operador de índice !! para acessar o mapa e determinar o terreno na posição fornecida. - As coordenadas x e y são arredondadas para inteiros usando floor antes de serem usadas como "índices".

processaInimigos

Arguments

:: [Inimigo]
Recebe a lista de inimigos a serem processados

-> Base
Recebe a base do jogo

-> ([Inimigo], Base)
Devolve a lista de inimigos restates e a base atualizada

A função "processaInimigos" é responsável por processar a lista de inimigos, atualizando o estado da base e r emovendo inimigos que tenham sido derrotados ou atingido a base.

Configuração:

Expand

A função "processaInimigos" utiliza a função "foldl" para processar cada inimigo da lista, verificando as seguintes condições: - Se o inimigo foi derrotado (vida <= 0), ele é removido e o butim do inimigo é adicionado aos créditos da base. - Se o inimigo atingiu a base, a vida da base é reduzida pelo ataque do inimigo. - Caso contrário, o inimigo permanece na lista.

atingiuBase

Arguments

:: Inimigo
Recebe o inimigo a ser verificado

-> Base
Recebe a base do jogador

-> Bool
Devolve um Booleano

A função "atingiuBase" verifica se um inimigo atingiu a base, comparando suas posições.

aplicarDanoFogo

Arguments

:: Tempo
Recebe o tempo decorrido desde a última atualização

-> [Projetil]
Recebe a lista de projeteis atualizados

-> Float
Devolve o dano a aplicar no devido frame

A função "aplicarDanoFogo" aplica dano aos inimigos com base nos projéteis de fogo que atingem o inimigo, levando em consideração a duração desses projéteis.

Funcionamento:

Expand

A função percorre a lista de projéteis e, caso o tipo do projétil seja Fogo, aplica um dano de 10 por segundo multiplicado pelo tempo. Se o projétil não for de fogo, não causa dano.

atualizaPortal

Arguments

:: Tempo
Recebe o Tempo decorrido do Jogo

-> Portal
Recebe um Portal do Jogo

-> Portal
Devolve o Portal com os respetivos valores atualizados

A função "atualizaPortal" é responsável por atualizar os temporizadores das ondas dos portais de acordo com o tempo decorrido no jogo. Ela integra a função "atualizaOnda", que atualiza os detalhes de cada onda individualmente.

Funcionamento:

A função percorre as ondas do portal e para cada onda chama a função "atualizaOnda" para atualizar o seu temporizador com base no tempo decorrido.

atualizaOnda

Arguments

:: Tempo
Recebe o Tempo decorrido do Jogo

-> Onda
Recebe uma Onda de um Portal

-> Onda
Devolve a Onda com os respetivos valores atualizados

A função "atualizaOnda" é responsável por atualizar o estado de uma onda de inimigos em um portal com base no tempo decorrido. Ela ajusta os temporizadores de ativação e lançamento dos inimigos de acordo com o tempo passado.

Comportamento:

Se a onda ainda não está ativa (entrada > 0), o tempo de entrada diminui.

Se a onda já está ativa (entrada == 0) e o delay de lançamento dos inimigos ainda não expirou (delay > 0), o tempo do delay diminui.

Se a onda está ativa e o delay expirou (delay == 0), o temporizador é reiniciado para o valor do ciclo de onda.

Funcionamento:

A função realiza a atualização dos campos da Onda conforme a condição: 1. Se a onda ainda não foi ativada, o tempo de entrada diminui. 2. Se a onda está ativa, mas ainda há delay, o tempo de delay diminui. 3. Se o delay expirou, o temporizador é resetado para o ciclo da onda.

ativaPortal

Arguments

:: Portal
Recebe um Portal do Jogo

-> (Portal, [Inimigo])
Devolve um par com o respetivo Portal atualizada, e a lista de Inimigo lançados por ele

A função "ativaPortal" é responsável por ativar um portal e lançar inimigos caso haja alguma onda ativa. Ela verifica se o portal possui ondas ativas e, se for o caso, chama a função ativaInimigo para gerar os inimigos correspondentes.

Comportamento:

Se o portal tiver uma onda ativa, a função chama "ativaInimigo" para gerar os inimigos e atualiza o portal.

Se o portal não tiver ondas ativas, ele permanece no mesmo estado e não lança inimigos.

Funcionamento:

A função verifica as ondas ativas no portal: - Se houver uma onda ativa (com "entradaOnda" maior que zero), a função chama "ativaInimigo" para lançar inimigos e atualiza o portal. - Se não houver ondas ativas, o portal permanece inalterado e não há inimigos gerados.

shouldSpawn

Arguments

:: Onda
Recebe uma Onda de um Portal

-> Bool
Devolve um Booleano

A função "shouldSpawn" verifica se uma onda deve ou não lançar um inimigo. A condição para o lançamento de um inimigo é que a onda esteja ativa e que o tempoOnda (tempo de delay) seja igual a 0.

Comportamento:

A onda é considerada ativa quando "entradaOnda" é 0, o que significa que a onda já foi ativada.

O lançamento de inimigos é condicionado a o tempo de delay ("tempoOnda") ser igual a 0, ou seja, não há mais atraso para o lançamento.

Funcionamento:

A função retorna True se a onda estiver ativa e sem delay, ou seja, pronta para lançar um inimigo. Caso contrário, retorna False.