Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2024/25.

Documentation

atualizaJogo :: Tempo -> Jogo -> Jogo

Atualiza o estado do jogo conforme o tempo passado.

Exemplos:

Expand

>>> atualizaJogo 1 (Jogo (Base 20 (5,2) 300) [Portal (0,0) [Onda [Inimigo (0,0) Oeste 10 1 1 1 [] 1] 2 2 0]] [] [[t,t],[r,t]] [Inimigo (1,0) Este 10 1 1 1 [] 1] [(100,torreResina),(150,torreFogo),(200,torreGelo)])
Jogo (Base 20 (5,2) 300) [Portal (0,0) [Onda [Inimigo (0,0) Oeste 10 1 1 1 [] 1] 2 1 0]] [] [[Terra,Terra],[Relva,Terra]] [Inimigo (1,1 ) Sul 10 1 1 1 [] 1] [(100,torreResina),(150,torreFogo),(200,torreGelo)]

atualizaTorre :: Tempo -> Torre -> [Inimigo] -> (Torre, [Inimigo])

Atualiza uma unica torre e os inimigos ativos no jogo.

atualizaTorres :: Tempo -> [Torre] -> [Inimigo] -> ([Torre], [Inimigo])

Atualiza todas as torres do jogo e os inimigos ativos no jogo.

Exemplos:

Expand

>>> atualizaTorres 1 [Torre (1,1) 1 1 1 1 0 (Projetil Fogo (Finita 10))] [Inimigo (1,1) Sul 1 1 1 1 [Projetil Fogo (Finita 10)] 1]
([Torre (1,1) 1 1 1 1 1 (Projetil Fogo (Finita 10))],[Inimigo (1,1) Sul 0 1 1 1 [Projetil Fogo (Finita 20)] 1]) 

inimigosVivoseCreditos :: [Inimigo] -> Creditos -> ([Inimigo], Creditos)

Retira os inimigos mortos da lista de inimigos.

atualizaPosicaoInimigo :: Tempo -> Inimigo -> Mapa -> Inimigo

Atualiza a posição de um inimigo com base na sua direção e velocidade.

atualizaInimigos :: Tempo -> [Inimigo] -> Mapa -> [Inimigo]

Atualiza a lista de inimigos no jogo.

Exemplos:

Expand

>>> atualizaInimigos 1 [Inimigo (1,0) Este 1 1 1 1 [Projetil Fogo (Finita 10)] 1] [[Terra,Terra],[Relva,Terra]]
[Inimigo (1,1) Sul 0.75 1 1 1 [Projetil Fogo (Finita 9)] 1] 

projetilInimigo :: Tempo -> Inimigo -> Inimigo

Aplica ao inimigo os efeitos dos projeteis em que nele estao ativos.

projeteisAtivos :: Tempo -> [Projetil] -> [Projetil]

Exclui da lista de uma lista de projeteis os projeteis que ja acabaram a sua duracao.

menosDuracao :: Duracao -> Tempo -> Duracao

Recebe uma duração e um tempo e devolve a duração menos o tempo.

inimigoBase :: [Inimigo] -> Base -> (Base, [Inimigo])

Recebe a lista de inimigos ativos e a base e devolve a um tuplo da base retirando a vida consuante se tem inimigos na mesma posicao e a lista de inimigos que estao numa posicao diferente da base.

baseNova :: [Inimigo] -> Base -> Base

Recebe a lista de inimigos ativos e a base e devolve a base retirando a vida consuante se tem inimigos na mesma posicao.

inimigosFiltrados :: [Inimigo] -> Posicao -> [Inimigo]

Recebe a lista de inimigos ativos e uma posicao e devolve a lista de inimigos que nao estao nessa posicao.

ondasFinal :: [Onda] -> [Onda]

Mantem so as ondas que ainda tem inimigos.

atualizaOnda :: Tempo -> Onda -> [Inimigo] -> (Onda, [Inimigo])

Atualiza uma onde conforme o seu tempo de entrada e de lançamento.

atualizaOndas :: Tempo -> [Onda] -> [Inimigo] -> ([Onda], [Inimigo])

Atualiza todas as ondas do jogo e os inimigos ativos no jogo.

Exemplos:

Expand

>>> atualizaOndas 1 [Onda [Inimigo (1,1) Sul 1 1 1 1 [] 1] 0 0 0] []
([],[Inimigo (1,1) Sul 1 1 1 1 [] 1])

atualizaPortal :: Tempo -> Portal -> [Inimigo] -> (Portal, [Inimigo])

Atualiza o estado de um portal e os inimigos no jogo.

atualizaPortais :: Tempo -> [Portal] -> [Inimigo] -> ([Portal], [Inimigo])

Atualiza todos os portais do jogo e os inimigos ativos no jogo.

Exemplos:

Expand

>>> tualizaPortais 1 [Portal (1,2) [Onda [Inimigo (1,1) Sul 1 1 1 1 [] 1] 1 0 0]] []
([Portal (1,2) []],[Inimigo (1,1) Sul 1 1 1 1 [] 1]) 

elementoDaMatriz :: (Float, Float) -> Mapa -> Maybe Terreno

Função que devolve o terreno da matriz na posição (linha,coluna).